% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_stress.R
\name{ac_calcStress}
\alias{ac_calcStress}
\title{Calculate map stress from coordinates}
\usage{
ac_calcStress(ag_coords, sr_coords, titer_table, colbases)
}
\arguments{
\item{ag_coords}{Matrix of antigen coordinates}

\item{sr_coords}{Matrix of sera coordinates}

\item{titer_table}{The titer table}

\item{colbases}{Titer table column bases}
}
\value{
Returns the total stress
}
\description{
Calculate map stress from coordinates
}
