% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_stress.R
\name{ac_calculate_stress}
\alias{ac_calculate_stress}
\title{Calculate map stress}
\usage{
ac_calculate_stress(map_dist, table_dist, lessthans, morethans)
}
\arguments{
\item{map_dist}{Map distance information}

\item{table_dist}{Table distance information}
}
\description{
Calculate stress from table and map distances.
}
