% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_stress.R
\name{ac_getTableColbases}
\alias{ac_getTableColbases}
\title{Get column bases from HI table}
\usage{
ac_getTableColbases(titer_table, minimum_column_basis = "none")
}
\arguments{
\item{titer_table}{The titer table}

\item{minimum_column_basis}{The minimum column basis to assume}
}
\value{
Returns the column bases as a numeric vector
}
\description{
Get column bases from HI table
}
