% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_stress.R
\name{ac_tableDists}
\alias{ac_tableDists}
\title{Calculate table distances}
\usage{
ac_tableDists(titer_table, colbases)
}
\arguments{
\item{titer_table}{The HI table.}

\item{colbases}{Column bases to use for each sera.}
}
\description{
Calculate table distances
}
