% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acmacs_placeAbvs.R
\name{acmacs_placeAbvs}
\alias{acmacs_placeAbvs}
\title{All known two-letter place name abbreviations}
\usage{
acmacs_placeAbvs()
}
\value{
Returns a matrix of known two-letter place name abbreviations
}
\description{
All known two-letter place name abbreviations
}
