% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{acmapAttributes}
\alias{acmapAttributes}
\alias{numAntigens}
\alias{numSera}
\alias{numSeraGroups}
\alias{numPoints}
\alias{numOptimizations}
\alias{numLayers}
\title{Get acmap attributes}
\usage{
numAntigens(map)

numSera(map)

numSeraGroups(map)

numPoints(map)

numOptimizations(map)

numLayers(map)
}
\arguments{
\item{map}{The acmap data object}
}
\value{
A number relating to the attribute
}
\description{
Functions to get various attributes about an acmap object.
}
\seealso{
Other map attribute functions:
\code{\link[=adjustedLogTiterTable]{adjustedLogTiterTable()}},
\code{\link[=adjustedTiterTable]{adjustedTiterTable()}},
\code{\link[=dilutionStepsize]{dilutionStepsize()}},
\code{\link[=logtiterTableLayers]{logtiterTableLayers()}},
\code{\link[=mapDescription]{mapDescription()}},
\code{\link[=mapName]{mapName()}},
\code{\link[=titerTable]{titerTable()}},
\code{\link[=titerTableFlat]{titerTableFlat()}},
\code{\link[=titerTableLayers]{titerTableLayers()}}
}
\concept{map attribute functions}
