% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimizations.R
\name{acmapCoords}
\alias{acmapCoords}
\alias{agCoords}
\alias{srCoords}
\alias{agCoords<-}
\alias{srCoords<-}
\title{Get and set acmap point coordinates}
\usage{
agCoords(map, optimization_number = NULL, names = TRUE)

srCoords(map, optimization_number = NULL, names = TRUE)

agCoords(map, optimization_number = NULL) <- value

srCoords(map, optimization_number = NULL) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization number to access (default is the currently selected optimization)}

\item{names}{Should coordinate rownames be labelled with antigen and serum names}
}
\description{
Functions to get and set antigen and serum coordinates for an acmap object. This
will apply to the currently selected optimization (see \code{\link{selectedOptimization}}) by default,
but a specific optimization number can also be passed as an argument.
}
\seealso{
Other functions to get and set acmap optimization attributes: \code{\link{mapComments}},
  \code{\link{mapDimensions}}, \code{\link{mapStress}},
  \code{\link{mapTransformation}},
  \code{\link{minColBasis}}
}
\concept{functions to get and set acmap optimization attributes}
