% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimize.R
\name{add_hemispheringData}
\alias{add_hemispheringData}
\title{Add data on hemisphering diagnostics to a map}
\usage{
add_hemispheringData(map, data = NULL, optimization_number = NULL)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The map optimization number (defaults to the currently selected optimization)}
}
\value{
Returns the map data with additional diagnostic information on hemisphering points included.
}
\description{
Add data on hemisphering diagnostics to a map
}
