% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_procrustes.R
\name{add_procrustesData}
\alias{add_procrustesData}
\title{Add procrustes data to a map}
\usage{
add_procrustesData(map, procrustes_data)
}
\arguments{
\item{map}{The acmap data object}

\item{procrustes_data}{The procrustes data}
}
\value{
Returns the map data with additional diagnostic information on hemisphering points included.
}
\description{
Add procrustes data to a map
}
