% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimizations.R
\name{addOptimization}
\alias{addOptimization}
\title{Add a new optimization to an acmap object}
\usage{
addOptimization(
  map,
  ag_coords = NULL,
  sr_coords = NULL,
  number_of_dimensions,
  minimum_column_basis = NULL,
  warnings = TRUE,
  ...
)
}
\arguments{
\item{map}{The acmap data object}

\item{ag_coords}{Antigen coordinates for the new optimization (random if not specified)}

\item{sr_coords}{Sera coordinates for the new optimization (random if not specified)}

\item{number_of_dimensions}{The number of dimensions of the new optimization}

\item{minimum_column_basis}{The minimum column basis to use for the new optimization}

\item{...}{Further optimization parameters}
}
\value{
Returns the acmap data object with new optimization added (but not selected).
}
\description{
Function to add a new optimization to an acmap object, with specified values.
}
