% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reactivity_adjusts.R
\name{adjustedLogTiterTable}
\alias{adjustedLogTiterTable}
\title{Get the reactivity adjusted log titer table}
\usage{
adjustedLogTiterTable(map, optimization_number = 1)
}
\arguments{
\item{map}{An acmap object}

\item{optimization_number}{The optimization number from which
to take any antigen reactivity adjustments}
}
\value{
A numeric matrix of adjusted log titers.
}
\description{
Return the log titer table plus any antigen reactivity adjustments.
}
\seealso{
Other map attribute functions:
\code{\link{acmapAttributes}},
\code{\link[=adjustedTiterTable]{adjustedTiterTable()}},
\code{\link[=dilutionStepsize]{dilutionStepsize()}},
\code{\link[=logtiterTableLayers]{logtiterTableLayers()}},
\code{\link[=mapDescription]{mapDescription()}},
\code{\link[=mapName]{mapName()}},
\code{\link[=titerTable]{titerTable()}},
\code{\link[=titerTableFlat]{titerTableFlat()}},
\code{\link[=titerTableLayers]{titerTableLayers()}}
}
\concept{map attribute functions}
