% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reactivity_adjusts.R
\name{adjustedTiterTable}
\alias{adjustedTiterTable}
\title{Get the reactivity adjusted titer table}
\usage{
adjustedTiterTable(map, optimization_number = 1)
}
\arguments{
\item{map}{An acmap object}

\item{optimization_number}{The optimization number from which
to take any antigen reactivity adjustments}
}
\value{
A character matrix of titers.
}
\description{
Return the titer table plus any antigen reactivity adjustments.
}
\seealso{
\code{\link[=htmlAdjustedTiterTable]{htmlAdjustedTiterTable()}}

Other map attribute functions:
\code{\link{acmapAttributes}},
\code{\link[=adjustedLogTiterTable]{adjustedLogTiterTable()}},
\code{\link[=dilutionStepsize]{dilutionStepsize()}},
\code{\link[=logtiterTableLayers]{logtiterTableLayers()}},
\code{\link[=mapDescription]{mapDescription()}},
\code{\link[=mapName]{mapName()}},
\code{\link[=titerTable]{titerTable()}},
\code{\link[=titerTableFlat]{titerTableFlat()}},
\code{\link[=titerTableLayers]{titerTableLayers()}}
}
\concept{map attribute functions}
