% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_sera.R
\name{agHomologousSr}
\alias{agHomologousSr}
\title{Get homologous sera for each antigen}
\usage{
agHomologousSr(map)
}
\arguments{
\item{map}{An acmap object}
}
\value{
A list, where each entry is a vector of indices for homologous
sera, or a length 0 vector where no homologous serum is present
}
\description{
Gets the indices of homologous sera for each antigen in an antigenic map.
See also the function \code{srHomologousAgs()} for getting and setting the
homologous antigens reciprocally.
}
\seealso{
Other antigen and sera attribute functions: 
\code{\link{agAttributes}},
\code{\link{agGroups}()},
\code{\link{agLabIDs}()},
\code{\link{agSequences}()},
\code{\link{ptAnnotations}},
\code{\link{ptClades}},
\code{\link{srAttributes}},
\code{\link{srGroups}()},
\code{\link{srHomologousAgs}()},
\code{\link{srSequences}()}
}
\concept{antigen and sera attribute functions}
