% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_antigens.R
\name{agLabIDs}
\alias{agLabIDs}
\alias{agLabIDs<-}
\title{Getting and setting antigen lab id information}
\usage{
agLabIDs(map)

agLabIDs(map) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{value}{A list of character vectors with lab ids information for each
point}
}
\value{
A character vector of antigen laboratory IDs
}
\description{
Getting and setting antigen lab id information
}
\seealso{
Other antigen and sera attribute functions: 
\code{\link{agAttributes}},
\code{\link{agGroups}()},
\code{\link{agHomologousSr}()},
\code{\link{agSequences}()},
\code{\link{ptAnnotations}},
\code{\link{ptClades}},
\code{\link{srAttributes}},
\code{\link{srGroups}()},
\code{\link{srHomologousAgs}()},
\code{\link{srSequences}()}
}
\concept{antigen and sera attribute functions}
