% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{agReactivityAdjustments}
\alias{agReactivityAdjustments}
\alias{agReactivityAdjustments<-}
\title{Get and set antigen reactivity adjustments}
\usage{
agReactivityAdjustments(map)

agReactivityAdjustments(map) <- value
}
\arguments{
\item{map}{The acmap object}

\item{value}{A vector of antigen reactivity adjustments to apply}
}
\value{
A numeric vector of antigen reactivity adjustments
}
\description{
Get and set antigen reactivity adjustments
}
\seealso{
Other functions for working with map data:
\code{\link[=acmap]{acmap()}},
\code{\link[=addOptimization]{addOptimization()}},
\code{\link[=as.json]{as.json()}},
\code{\link[=edit_agNames]{edit_agNames()}},
\code{\link[=edit_srNames]{edit_srNames()}},
\code{\link[=keepBestOptimization]{keepBestOptimization()}},
\code{\link[=keepSingleOptimization]{keepSingleOptimization()}},
\code{\link[=layerNames]{layerNames()}},
\code{\link{orderPoints}},
\code{\link[=read.acmap]{read.acmap()}},
\code{\link[=read.titerTable]{read.titerTable()}},
\code{\link{removePoints}},
\code{\link[=save.acmap]{save.acmap()}},
\code{\link[=save.coords]{save.coords()}},
\code{\link[=save.titerTable]{save.titerTable()}},
\code{\link{subsetCommonPoints}},
\code{\link[=subsetMap]{subsetMap()}}
}
\concept{functions for working with map data}
