% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_procrustes.R
\name{apply_procrustes}
\alias{apply_procrustes}
\title{Apply a procrustes object}
\usage{
apply_procrustes(coords, pc_object, dim_match = TRUE,
  rotate2Dto3D = FALSE)
}
\arguments{
\item{coords}{Coordinates to be rotated.}

\item{pc_object}{Procrustes object to be applied.}
}
\value{
Returns the coordinates with the procrustes transformation applied
}
\description{
Apply a procrustes object
}
