% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_transformation.R
\name{applyMapTransform}
\alias{applyMapTransform}
\title{Apply the current map transformation}
\usage{
applyMapTransform(coords, map, optimization_number = NULL)
}
\arguments{
\item{coords}{Coordinates to transform}

\item{map}{The acmap object}

\item{optimization_number}{The optimization number (by default the selected one)}
}
\description{
Applies the current map transformation to a set of coordinates
}
\seealso{
Other {functions relating to map transformation}: 
\code{\link{reflectMap}()},
\code{\link{rotateMap}()},
\code{\link{translateMap}()}
}
\concept{{functions relating to map transformation}}
