% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plotspec.R
\name{applyPlotspec}
\alias{applyPlotspec}
\title{Apply a plotspec from another acmap}
\usage{
applyPlotspec(map, source_map)
}
\arguments{
\item{map}{The acmap object}

\item{source_map}{An acmap object from which to copy point styles}
}
\value{
Returns the acmap object with updated point styles (unmatched point
styles unchanged)
}
\description{
Copy point style from matching antigens and sera in another acmap
}
\seealso{
Other map point style functions: 
\code{\link{ptDrawingOrder}()},
\code{\link{ptOpacity}},
\code{\link{ptStyles}}
}
\concept{map point style functions}
