% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.R
\name{blob}
\alias{blob}
\title{Plot a blob object}
\usage{
blob(x, col, border, lwd, alpha = 1, ...)
}
\arguments{
\item{x}{The blob object to plot}

\item{col}{Color for the blob fill}

\item{border}{Color for the blob outline}

\item{lwd}{Line width for the blob outline}

\item{alpha}{Blob opacity}

\item{...}{Additional arguments to pass to \code{polygon()}}
}
\value{
No return value, called for the side effect of plotting the blobs.
}
\description{
Plot a blob object such as that return from \code{agBootstrapBlob()} using the
\code{polygon()} function.
}
\seealso{
Other additional plotting functions: 
\code{\link{blobsize}()}
}
\concept{additional plotting functions}
