% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_blobs.R
\name{blobsize}
\alias{blobsize}
\title{Calculate size of a blob object}
\usage{
blobsize(blob)
}
\arguments{
\item{blob}{The blob object}
}
\value{
A numeric vector
}
\description{
Returns either the area (for 2D blobs) or volume (for 3D blobs)
}
\seealso{
Other additional plotting functions: 
\code{\link{blob}()}
}
\concept{additional plotting functions}
