% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_diagnostics_bootstrap.R
\name{bootstrapMap}
\alias{bootstrapMap}
\title{Perform a noisy bootstrap on a map}
\usage{
bootstrapMap(
  map,
  bootstrap_repeats = 1000,
  optimizations_per_repeat = 100,
  ag_noise_sd = 0.7,
  titer_noise_sd = 0.7,
  options = list()
)
}
\arguments{
\item{map}{The map object}

\item{bootstrap_repeats}{The number of bootstrap repeats to perform}

\item{optimizations_per_repeat}{When reoptimizing the map from scratch, the
number of optimization runs to perform}

\item{ag_noise_sd}{The standard deviation (on the log titer scale) to use
when applying noise per antigen}

\item{titer_noise_sd}{The standard deviation (on the log titer scale) to use
when applying noise per titer}

\item{options}{Map optimizer options, see \code{RacOptimizer.options()}}
}
\value{
Returns the map object updated with bootstrap information
}
\description{
This function takes the map and original titer table, converts the titers to
the log scale and reoptimizes the map from scratch. For each bootstrap run
this process is performed and a record of the coordinates of points in the
lowest stress solution is kept. This can only be performed on a map that
has already been optimized, the intention is that you can use this function
to see how robust point positions are in that map when additional noise is
simulated.
}
\seealso{
Other {map diagnostic functions}: 
\code{\link{bootstrapBlobs}()},
\code{\link{checkHemisphering}()},
\code{\link{dimensionTestMap}()},
\code{\link{mapRelaxed}()}
}
\concept{{map diagnostic functions}}
