% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_coord_dist}
\alias{calc_coord_dist}
\title{Calculate the distance between two sets of coordinates}
\usage{
calc_coord_dist(coords1, coords2)
}
\description{
This calculates the euclidean distance, row by row, of two sets of coordinates.
}
