% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_procrustes.R
\name{calc_procrustes}
\alias{calc_procrustes}
\title{Calculate a procrustes object}
\usage{
calc_procrustes(source_coords, target_coords, translation = TRUE,
  scaling = FALSE, dim_match = TRUE)
}
\arguments{
\item{source_coords}{Source coordinates.}

\item{target_coords}{Target coordinates.}

\item{translation}{Should translation be applied?}

\item{scaling}{Should scaling be applied?}

\item{dim_match}{Should automatic dimension matching be done?}
}
\value{
Returns a procrustes object.
}
\description{
Calculate a procrustes object
}
