% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_titer_lims.R
\name{calc_titer_lims}
\alias{calc_titer_lims}
\title{Get titer limits}
\usage{
calc_titer_lims(titers, min_titer_possible = -Inf,
  max_titer_possible = Inf)
}
\arguments{
\item{titers}{A numeric/character vector or matrix of titer measurements.}

\item{min_titer_possible}{The maximum possible titer to assume}

\item{max_titer_possible}{The minimum possible titer to assume}
}
\value{
Returns a list of length two with values max_titers and min_titers, giving the
numeric vectors of the upper and lower bounds of the titers on the log scale.
}
\description{
Function for getting upper and lower limits of measured titers on the log scale.
}
\details{
This function assumes that HI measurements were performed in 2-fold dilution
steps and converts them to the log scale using the formula:
\deqn{a + b}
Hence an HI titer of 20, which would convert to 1 via the transformation above, would be
assumed to have upper and lower limits of 1.5 and 0.5 respectively.

In the case of non-detectable titers, such as <10, the lower bound of the measured value
is taken from the parameter \code{min_titer_possible}, defaulting to the value found from
a call to \code{get_lndscp_fit_defaults()}. For a greater than value, i.e. >1280, the
upper bound of the value is taken from the parameter \code{max_titer_possible}. You can
set different defaults by passing them as named arguments to the list, as shown in the
examples.
}
\examples{
# Calculate the titer limits of a set of HI titers
titer_lims <- get_titer_lims(titers = c("20", "320", "<10", ">1280"))

# Calculate the titer limits assuming non-default upper and lower bounds for non-detectable
# and greater-than titers.
titer_lims <- get_titer_lims(titers = c("20", "320", "<10", ">1280"),
                             fit_opts = list(min_titer_possible = -Inf,
                                             max_titer_possible = 14))
}
