% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_blobs.R
\name{calculate_stressBlob}
\alias{calculate_stressBlob}
\title{Calculate stress blob information}
\usage{
calculate_stressBlob(map, optimization_number = NULL, stress_lim = 1,
  antigens = TRUE, sera = TRUE, grid_spacing = 0.1,
  grid_margin = 4, progress_fn = message)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization number (defaults to the currently selected one)}

\item{stress_lim}{The blob stress limit}

\item{antigens}{Antigens to calculate blobs for (TRUE for all FALSE for none, or specified by name or index)}

\item{sera}{Sera to calculate blobs for (TRUE for all FALSE for none, or specified by name or index)}

\item{grid_spacing}{Grid spacing to use when calculating the blob}

\item{grid_margin}{Grid margin to use when calculating the blob}

\item{progress_fn}{Function to use for progress reporting}
}
\value{
Returns stress blob information
}
\description{
Calculate stress blob information
}
