% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimize.R
\name{checkHemisphering}
\alias{checkHemisphering}
\title{Check for hemisphering or trapped points}
\usage{
checkHemisphering(map, optimization_number = NULL)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The map optimization number (defaults to the currently selected optimization)}
}
\value{
Returns a data frame with information on any points that were found
  to be hemisphering or trapped.
}
\description{
Check for hemisphering or trapped points
}
