% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_new.R
\name{cloneMap}
\alias{cloneMap}
\title{Clone an acmap object}
\usage{
cloneMap(map)
}
\arguments{
\item{map}{The map object}
}
\value{
Returns a copy of the map object
}
\description{
Creates a copy of an acmap object. This is needed when working with
the acmap.cpp map types since these are environments rather than lists
and are not copy-on-modify, i.e. if you change one reference to the
acmap, all references to it will be changed. To avoid this behaviour
you can use the cloneMap function.
}
\seealso{
Other {functions for working with map data}: 
\code{\link{acmap}()},
\code{\link{as.json}()},
\code{\link{convertingMaps}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{save.acmap}()},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{selectedOptimization}()},
\code{\link{subsetMap}()}
}
\concept{{functions for working with map data}}
