% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimizations.R
\name{colBases}
\alias{colBases}
\title{Get the column basis for an acmap}
\usage{
colBases(map, optimization_number = NULL, name = TRUE)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization data to access (defaults to the currently selected optimization)}
}
\value{
Returns a numeric vector of the column bases used for this map.
}
\description{
Get the numeric column basis associated with the currently selected, or specified acmap optimization.
Once a optimization has been created the column bases cannot be further modified but instead a new
optimization must be created (see \code{\link{addOptimization}}).
}
\seealso{
\code{\link{minColBasis}} For getting the minimum column basis that was set (or to know if column bases were fixed).
}
