% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_utils_manipulate_titers.R
\name{convert2log}
\alias{convert2log}
\title{Convert raw titers to log titers}
\usage{
convert2log(titers)
}
\arguments{
\item{titers}{The raw titers to convert.}
}
\description{
Convert raw titers to log titers
}
