% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_utils_manipulate_titers.R
\name{convert2raw}
\alias{convert2raw}
\title{Function for converting from log titers to raw titers}
\usage{
convert2raw(titers, lessthan_titers, morethan_titers,
  autoconvert_lessthans = missing(lessthan_titers))
}
\arguments{
\item{titers}{The titers}

\item{titer_type}{The titer type one of ("disc", "lessthan" or "morethan")}
}
\description{
Function for converting from log titers to raw titers
}
