% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_conversion.R, R/map_new.R
\name{convertingMaps}
\alias{convertingMaps}
\alias{as.list.racchart}
\alias{as.cpp}
\alias{acmap.cpp}
\title{Converting between map formats}
\usage{
\method{as.list}{racchart}(map)

as.cpp(map)

acmap.cpp(...)
}
\arguments{
\item{map}{The map object to be converted.}
}
\value{
Returns the converted map data object.
}
\description{
Functions to convert between the 'racmap' and 'racchart' formats (see details).
}
\details{
There are two map data formats supported within Racmacs termed an
'acmap' and an 'acchart'.

In short, if all you want to do is access the
map data for your own plotting or visualization or analysis, you can
choose the 'acmap' format, which is simply an R \code{\link[base]{list}}.
}
\seealso{
Other {functions for working with map data}: 
\code{\link{acmap}()},
\code{\link{as.json}()},
\code{\link{cloneMap}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{removePoints}},
\code{\link{save.acmap}()},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{selectedOptimization}()},
\code{\link{subsetMap}()},
\code{\link{subsetPoints}}

Other {functions for working with map data}: 
\code{\link{acmap}()},
\code{\link{as.json}()},
\code{\link{cloneMap}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{removePoints}},
\code{\link{save.acmap}()},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{selectedOptimization}()},
\code{\link{subsetMap}()},
\code{\link{subsetPoints}}

Other {functions for working with map data}: 
\code{\link{acmap}()},
\code{\link{as.json}()},
\code{\link{cloneMap}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{removePoints}},
\code{\link{save.acmap}()},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{selectedOptimization}()},
\code{\link{subsetMap}()},
\code{\link{subsetPoints}}
}
\concept{{functions for working with map data}}
