% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{CrossProduct3D}
\alias{CrossProduct3D}
\title{Function to get the 3D cross product of two vectors}
\usage{
CrossProduct3D(U, V)
}
\arguments{
\item{U}{Vector 1}

\item{V}{Vector 2}
}
\value{
Returns the cross product
}
\description{
Function to get the 3D cross product of two vectors
}
