% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_blobs.R
\name{deprecated_fn}
\alias{deprecated_fn}
\alias{stressBlobs}
\alias{agStressBlobSize}
\alias{srStressBlobSize}
\title{Deprecated functions}
\usage{
stressBlobs(...)

agStressBlobSize(...)

srStressBlobSize(...)
}
\arguments{
\item{...}{Arguments to pass to the new function}
}
\description{
These functions still work but have been deprecated in favour of another function. Arguments will be passed onto the new function with a warning.
}
