% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Racmacs.R
\name{.onLoad}
\alias{.onLoad}
\title{Hooks for setting package options on load}
\usage{
.onLoad(libname, pkgname)
}
\description{
Hooks for setting package options on load
}
\details{
\code{Racmacs.parallel}:
Should optimizations be run in parallel. If true
this will speed up computation, but can sometimes lead to instability.
}
