% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_edit_names.R
\name{edit_agNames}
\alias{edit_agNames}
\title{Edit antigen names in an acmap}
\usage{
edit_agNames(map, old_names, new_names)
}
\arguments{
\item{map}{The map data object to be updated}

\item{old_names}{Old names to be replaced}

\item{new_names}{Replacement for old names}
}
\value{
Returns the acmap object with antigen names updated.
}
\description{
Edit antigen names in an acmap
}
