% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_edit_names.R
\name{edit_srNames}
\alias{edit_srNames}
\title{Edit sera names in an acmap}
\usage{
edit_srNames(map, old_names, new_names)
}
\arguments{
\item{map}{The map data object to be updated}

\item{old_names}{Old names to be replaced}

\item{new_names}{Replacement for old names}
}
\value{
Returns the acmap object with sera names updated.
}
\description{
Edit sera names in an acmap
}
\seealso{
Other functions for working with map data:
\code{\link[=acmap]{acmap()}},
\code{\link[=addOptimization]{addOptimization()}},
\code{\link[=agReactivityAdjustments]{agReactivityAdjustments()}},
\code{\link[=as.json]{as.json()}},
\code{\link[=edit_agNames]{edit_agNames()}},
\code{\link[=keepBestOptimization]{keepBestOptimization()}},
\code{\link[=keepSingleOptimization]{keepSingleOptimization()}},
\code{\link[=layerNames]{layerNames()}},
\code{\link{orderPoints}},
\code{\link[=read.acmap]{read.acmap()}},
\code{\link[=read.titerTable]{read.titerTable()}},
\code{\link{removePoints}},
\code{\link[=save.acmap]{save.acmap()}},
\code{\link[=save.coords]{save.coords()}},
\code{\link[=save.titerTable]{save.titerTable()}},
\code{\link{subsetCommonPoints}},
\code{\link[=subsetMap]{subsetMap()}}
}
\concept{functions for working with map data}
