% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_viewer.R
\name{export_viewer}
\alias{export_viewer}
\title{Export the map viewer}
\usage{
export_viewer(map, file, selfcontained = TRUE, ...)
}
\arguments{
\item{map}{The acmap object}

\item{file}{File to save HTML into}

\item{selfcontained}{Whether to save the HTML as a single self-contained file
(with external resources base64 encoded) or a file with external resources
placed in an adjacent directory.}

\item{...}{Further parameters to \code{view()}}
}
\description{
Export a map in a standalone html viewer
}
