% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_subset.R
\name{get_ag_indices}
\alias{get_ag_indices}
\title{Get indices of matching antigens in a map}
\usage{
get_ag_indices(antigens = TRUE, map, warnings = TRUE)
}
\arguments{
\item{antigens}{The antigens to get indices}

\item{map}{The acmap object}

\item{warnings}{Should warnings be output about antigens that were specified by name but not found in the map}
}
\value{
Returns the indices of matches antigens
}
\description{
Get indices of matching antigens in a map
}
