% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_subset.R
\name{get_sr_indices}
\alias{get_sr_indices}
\title{Get indices of matching sera in a map}
\usage{
get_sr_indices(sera = TRUE, map, warnings = TRUE)
}
\arguments{
\item{sera}{The sera to get indices}

\item{map}{The acmap object}

\item{warnings}{Should warnings be output about sera that were specified by name but not found in the map}
}
\value{
Returns the indices of matches sera
}
\description{
Get indices of matching sera in a map
}
