% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_viewer.R
\name{get_viewer_options}
\alias{get_viewer_options}
\title{Set viewer settings for a map}
\usage{
get_viewer_options(map)
}
\arguments{
\item{map}{}
}
\value{

}
\description{
A function to control visual features of the map display in the interactive viewer.
}
