% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot_gridplot.R
\name{grid.plot.acmap}
\alias{grid.plot.acmap}
\title{Plot an antigenic map using the grid system}
\usage{
grid.plot.acmap(
  map,
  optimization_number = 1,
  xlim = NULL,
  ylim = NULL,
  plot_ags = TRUE,
  plot_sr = TRUE,
  plot_labels = FALSE,
  plot_blobs = TRUE,
  grid.col = "grey90",
  grid.margin.col = grid.col,
  fill.alpha = 0.8,
  outline.alpha = 0.8,
  padding = 1,
  cex = 1
)
}
\description{
Plot an antigenic map using the 'grid' plotting system, this is harder to edit afterwards but
allows maps with custom point shapes like "EGG" and "UGLYEGG" to be plotted, or maps where
point aspect or rotation has been altered.
}
\seealso{
Other {functions to view maps}: 
\code{\link{plot.acmap}()},
\code{\link{snapshotMap}()}
}
\concept{{functions to view maps}}
