% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{keepOptimizations}
\alias{keepOptimizations}
\title{Keep specified optimization runs}
\usage{
keepOptimizations(map, optimization_numbers)
}
\arguments{
\item{map}{The acmap object}

\item{optimization_numbers}{Optimizations to keep}
}
\value{
Returns the updated acmap object
}
\description{
Keep only data from specified optimization runs.
}
\seealso{
Other functions to work with map optimizations: 
\code{\link{optimizationProperties}},
\code{\link{removeOptimizations}()},
\code{\link{sortOptimizations}()}
}
\concept{functions to work with map optimizations}
