% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimizations.R
\name{listOptimizations}
\alias{listOptimizations}
\title{Get all optimization details from an acmap object}
\usage{
listOptimizations(map)
}
\arguments{
\item{map}{The acmap data object}
}
\value{
Returns a list of lists with information about the optimizations
}
\description{
Gets the details associated with the all the optimizations of an acmap object
as a list.
}
\seealso{
See \code{getOptimization()} for getting information about a single
optimization.
}
