% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stress.R
\name{logtiterTable}
\alias{logtiterTable}
\title{Get the log titers from an acmap}
\usage{
logtiterTable(map)
}
\arguments{
\item{map}{The acmap object}
}
\value{
Returns a matrix of titers converted to the log scale
}
\description{
Converts titers to the log scale via via the transformation
$log2(x/10)$, lessthan values are reduced by 1 on the log scale and greater
than values are increased by 1, hence <10 => -1 and >1280 => 8
}
\seealso{
Other map diagnostic functions:
\code{\link[=agCohesion]{agCohesion()}},
\code{\link[=bootstrapBlobs]{bootstrapBlobs()}},
\code{\link[=bootstrapMap]{bootstrapMap()}},
\code{\link[=checkHemisphering]{checkHemisphering()}},
\code{\link[=dimensionTestMap]{dimensionTestMap()}},
\code{\link{map-table-distances}},
\code{\link{mapBootstrapCoords}},
\code{\link[=mapDistances]{mapDistances()}},
\code{\link[=mapRelaxed]{mapRelaxed()}},
\code{\link[=mapResiduals]{mapResiduals()}},
\code{\link{pointStress}},
\code{\link{ptBootstrapBlob}},
\code{\link[=ptBootstrapCoords]{ptBootstrapCoords()}},
\code{\link{ptLeverage}},
\code{\link{ptTriangulationBlob}},
\code{\link[=recalculateStress]{recalculateStress()}},
\code{\link[=stressTable]{stressTable()}},
\code{\link[=tableColbases]{tableColbases()}},
\code{\link[=tableDistances]{tableDistances()}},
\code{\link[=triangulationBlobs]{triangulationBlobs()}},
\code{\link{unstableMaps}}

Other functions relating to map stress calculation:
\code{\link[=mapDistances]{mapDistances()}},
\code{\link[=mapResiduals]{mapResiduals()}},
\code{\link{pointStress}},
\code{\link[=recalculateStress]{recalculateStress()}},
\code{\link[=stressTable]{stressTable()}},
\code{\link[=tableColbases]{tableColbases()}},
\code{\link[=tableDistances]{tableDistances()}}
}
\concept{functions relating to map stress calculation}
\concept{map diagnostic functions}
