% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RacViewer.R
\name{map2json}
\alias{map2json}
\title{Convert a map to json data}
\usage{
map2json(map)
}
\arguments{
\item{map}{The acmap data}
}
\value{
Returns json as a string
}
\description{
Convert a map to json data
}
