% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimizations.R
\name{mapComments}
\alias{mapComments}
\alias{mapComment}
\alias{mapComment<-}
\title{Get or set acmap comments}
\usage{
mapComment(map, optimization_number = NULL)

mapComment(map, optimization_number = NULL) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization data to access (defaults to the currently selected optimization)}
}
\description{
Get or set comments associated with a map optimization.
}
\seealso{
Other functions to get and set acmap optimization attributes: \code{\link{acmapCoords}},
  \code{\link{mapDimensions}}, \code{\link{mapStress}},
  \code{\link{mapTransformation}},
  \code{\link{minColBasis}}
}
\concept{functions to get and set acmap optimization attributes}
