% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_optimizations.R
\name{mapDimensions}
\alias{mapDimensions}
\title{Get the current map dimensions}
\usage{
mapDimensions(map, optimization_number = 1)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization run from which to get / set the data}
}
\value{
Returns the number of dimensions for the optimization run.
}
\description{
Get the current map dimensions
}
\seealso{
Other map optimization attribute functions:
\code{\link[=colBases]{colBases()}},
\code{\link[=mapComment]{mapComment()}},
\code{\link[=mapStress]{mapStress()}},
\code{\link[=mapTransformation]{mapTransformation()}},
\code{\link[=ptBaseCoords]{ptBaseCoords()}},
\code{\link[=ptCoords]{ptCoords()}}
}
\concept{map optimization attribute functions}
