% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stress.R
\name{mapDistances}
\alias{mapDistances}
\title{Return calculated map distances for an acmap}
\usage{
mapDistances(map, optimization_number = 1)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization number}
}
\value{
Returns a matrix of map distances with antigens as rows and sera as
columns.
}
\description{
Takes the acmap object and calculates euclidean distances between antigens
and sera for the currently selected or specified optimization.
}
\seealso{
Other map diagnostic functions:
\code{\link[=agCohesion]{agCohesion()}},
\code{\link[=bootstrapBlobs]{bootstrapBlobs()}},
\code{\link[=bootstrapMap]{bootstrapMap()}},
\code{\link[=checkHemisphering]{checkHemisphering()}},
\code{\link[=dimensionTestMap]{dimensionTestMap()}},
\code{\link[=logtiterTable]{logtiterTable()}},
\code{\link{map-table-distances}},
\code{\link{mapBootstrapCoords}},
\code{\link[=mapRelaxed]{mapRelaxed()}},
\code{\link[=mapResiduals]{mapResiduals()}},
\code{\link{pointStress}},
\code{\link{ptBootstrapBlob}},
\code{\link[=ptBootstrapCoords]{ptBootstrapCoords()}},
\code{\link{ptLeverage}},
\code{\link{ptTriangulationBlob}},
\code{\link[=recalculateStress]{recalculateStress()}},
\code{\link[=stressTable]{stressTable()}},
\code{\link[=tableColbases]{tableColbases()}},
\code{\link[=tableDistances]{tableDistances()}},
\code{\link[=triangulationBlobs]{triangulationBlobs()}},
\code{\link{unstableMaps}}

Other functions relating to map stress calculation:
\code{\link[=logtiterTable]{logtiterTable()}},
\code{\link[=mapResiduals]{mapResiduals()}},
\code{\link{pointStress}},
\code{\link[=recalculateStress]{recalculateStress()}},
\code{\link[=stressTable]{stressTable()}},
\code{\link[=tableColbases]{tableColbases()}},
\code{\link[=tableDistances]{tableDistances()}}
}
\concept{functions relating to map stress calculation}
\concept{map diagnostic functions}
