% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gadget.R
\name{mapGadget}
\alias{mapGadget}
\title{Open a shiny gadget to view the map}
\usage{
mapGadget(map)
}
\arguments{
\item{map}{The acmap object to open in the GUI}
}
\value{
No value returned, called for the side effect of starting the gadget.
}
\description{
This function is equivalent to running \code{runGUI()} and loading a map file, but this
takes the acmap object to open as an input argument.
}
\seealso{
Other functions to view maps:
\code{\link[=RacViewer]{RacViewer()}},
\code{\link[=RacViewer.options]{RacViewer.options()}},
\code{\link[=export_viewer]{export_viewer()}},
\code{\link[=ggplot.acmap]{ggplot.acmap()}},
\code{\link[=plot.acmap]{plot.acmap()}},
\code{\link[=setLegend]{setLegend()}},
\code{\link[=view]{view()}},
\code{\link[=view.acmap]{view.acmap()}},
\code{\link[=view.default]{view.default()}}
}
\concept{functions to view maps}
