% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{mapName}
\alias{mapName}
\alias{mapName<-}
\title{Getting and setting the map name}
\usage{
mapName(map)
mapName(map) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{value}{New value to set}
}
\value{
Returns either the requested attribute when using a getter function or

the updated acmap object when using the setter function.
}
\description{
Getting and setting the map name
}
\seealso{
Other map attribute functions:
\code{\link{acmapAttributes}},
\code{\link[=adjustedLogTiterTable]{adjustedLogTiterTable()}},
\code{\link[=adjustedTiterTable]{adjustedTiterTable()}},
\code{\link[=dilutionStepsize]{dilutionStepsize()}},
\code{\link[=logtiterTableLayers]{logtiterTableLayers()}},
\code{\link[=mapDescription]{mapDescription()}},
\code{\link[=titerTable]{titerTable()}},
\code{\link[=titerTableFlat]{titerTableFlat()}},
\code{\link[=titerTableLayers]{titerTableLayers()}}
}
\concept{map attribute functions}
