% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plotspec.R
\name{mapPlotspec}
\alias{mapPlotspec}
\alias{agFill}
\alias{srFill}
\alias{agFill<-}
\alias{srFill<-}
\alias{agOutline}
\alias{srOutline}
\alias{agOutline<-}
\alias{srOutline<-}
\alias{agOutlineWidth}
\alias{srOutlineWidth}
\alias{agOutlineWidth<-}
\alias{srOutlineWidth<-}
\alias{agSize}
\alias{srSize}
\alias{agSize<-}
\alias{srSize<-}
\alias{agRotation}
\alias{srRotation}
\alias{agRotation<-}
\alias{srRotation<-}
\alias{agAspect}
\alias{srAspect}
\alias{agAspect<-}
\alias{srAspect<-}
\alias{agShape}
\alias{srShape}
\alias{agShape<-}
\alias{srShape<-}
\alias{agShown}
\alias{srShown}
\alias{agShown<-}
\alias{srShown<-}
\alias{agDrawingOrder}
\alias{srDrawingOrder}
\alias{agDrawingOrder<-}
\alias{srDrawingOrder<-}
\alias{mapPoints}
\title{Setting point appearance in an antigenic map}
\usage{
agFill(map)

srFill(map)

agFill(map) <- value

srFill(map) <- value

agOutline(map)

srOutline(map)

agOutline(map) <- value

srOutline(map) <- value

agOutlineWidth(map)

srOutlineWidth(map)

agOutlineWidth(map) <- value

srOutlineWidth(map) <- value

agSize(map)

srSize(map)

agSize(map) <- value

srSize(map) <- value

agRotation(map)

srRotation(map)

agRotation(map) <- value

srRotation(map) <- value

agAspect(map)

srAspect(map)

agAspect(map) <- value

srAspect(map) <- value

agShape(map)

srShape(map)

agShape(map) <- value

srShape(map) <- value

agShown(map)

srShown(map)

agShown(map) <- value

srShown(map) <- value

agDrawingOrder(map)

srDrawingOrder(map)

agDrawingOrder(map) <- value

srDrawingOrder(map) <- value

mapPoints(map, optimization_number = NULL)
}
\arguments{
\item{map}{The map object}

\item{value}{Value to be used (recycled as necessary)}
}
\value{
Returns an updated racmap object
}
\description{
Setting point appearance in an antigenic map
}
