% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimize.R
\name{mapRelaxed}
\alias{mapRelaxed}
\title{Check if a map has been fully relaxed}
\usage{
mapRelaxed(map, optimization_number = 1, options = list())
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The map optimization number}

\item{options}{List of named optimizer options, see \code{RacOptimizer.options()}}
}
\value{
Returns TRUE or FALSE
}
\description{
Checks if the map optimization run can be relaxed further.
}
\seealso{
Other map diagnostic functions:
\code{\link[=agCohesion]{agCohesion()}},
\code{\link[=bootstrapBlobs]{bootstrapBlobs()}},
\code{\link[=bootstrapMap]{bootstrapMap()}},
\code{\link[=checkHemisphering]{checkHemisphering()}},
\code{\link[=dimensionTestMap]{dimensionTestMap()}},
\code{\link[=logtiterTable]{logtiterTable()}},
\code{\link{map-table-distances}},
\code{\link{mapBootstrapCoords}},
\code{\link[=mapDistances]{mapDistances()}},
\code{\link[=mapResiduals]{mapResiduals()}},
\code{\link{pointStress}},
\code{\link{ptBootstrapBlob}},
\code{\link[=ptBootstrapCoords]{ptBootstrapCoords()}},
\code{\link{ptLeverage}},
\code{\link{ptTriangulationBlob}},
\code{\link[=recalculateStress]{recalculateStress()}},
\code{\link[=stressTable]{stressTable()}},
\code{\link[=tableColbases]{tableColbases()}},
\code{\link[=tableDistances]{tableDistances()}},
\code{\link[=triangulationBlobs]{triangulationBlobs()}},
\code{\link{unstableMaps}}
}
\concept{map diagnostic functions}
