% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_antigens.R, R/map_sera.R
\name{mapStrains}
\alias{mapStrains}
\alias{agNames}
\alias{agNames<-}
\alias{agNamesFull}
\alias{agNamesFull<-}
\alias{agNamesAbbreviated}
\alias{agNamesAbbreviated<-}
\alias{agDates}
\alias{agDates<-}
\alias{agReference}
\alias{agReference<-}
\alias{srNames}
\alias{srNames<-}
\alias{srNamesFull}
\alias{srNamesFull<-}
\alias{srNamesAbbreviated}
\alias{srNamesAbbreviated<-}
\title{Strains in an antigenic map}
\usage{
agNames(map)

agNames(map) <- value

agNamesFull(map)

agNamesFull(map) <- value

agNamesAbbreviated(map)

agNamesAbbreviated(map) <- value

agDates(map)

agDates(map) <- value

agReference(map)

agReference(map) <- value

srNames(map)

srNames(map) <- value

srNamesFull(map)

srNamesFull(map) <- value

srNamesAbbreviated(map)

srNamesAbbreviated(map) <- value
}
\arguments{
\item{map}{The map object}

\item{value}{Value to be used (recycled as necessary)}

\item{map}{The map object}

\item{value}{Value to be used (recycled as necessary)}
}
\value{
Returns an updated racmap object

Returns an updated racmap object
}
\description{
Strains in an antigenic map

Strains in an antigenic map
}
