% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{mapSubtype}
\alias{mapSubtype}
\alias{mapSubtype<-}
\title{Getting and setting the map subtype}
\usage{
mapSubtype(map)
mapSubtype(map) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{value}{New value to set}
}
\value{
Returns either the requested attribute when using a getter function or

the updated acmap object when using the setter function.
}
\description{
Getting and setting the map subtype
}
\seealso{
Other map attribute functions: 
\code{\link{acmapAttributes}},
\code{\link{adjustedLogTiterTable}()},
\code{\link{adjustedTiterTable}()},
\code{\link{dilutionStepsize}()},
\code{\link{logtiterTableLayers}()},
\code{\link{mapAssay}()},
\code{\link{mapCreated}()},
\code{\link{mapDescription}()},
\code{\link{mapDetails}()},
\code{\link{mapLaboratory}()},
\code{\link{mapName}()},
\code{\link{mapRBCs}()},
\code{\link{mapType}()},
\code{\link{mapVirus}()},
\code{\link{titerTable}()},
\code{\link{titerTableFlat}()},
\code{\link{titerTableLayers}()}
}
\concept{map attribute functions}
