% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_transformation.R
\name{mapTransformation}
\alias{mapTransformation}
\alias{scaleMap}
\alias{transformMap}
\alias{translateMap}
\alias{reflectMap}
\alias{rotateMap}
\title{Apply transformations to an antigenic map}
\usage{
scaleMap(map, scaling, optimization_number = NULL)

transformMap(map, transformation_matrix, optimization_number = NULL)

translateMap(map, translation, optimization_number = NULL)

reflectMap(map, axis = "x", optimization_number = NULL)

rotateMap(map, degrees, axis = NULL, optimization_number = NULL)
}
\arguments{
\item{map}{The map object}

\item{value}{Value to be used (recycled as necessary)}
}
\value{
Returns an updated racmap object
}
\description{
Apply transformations to an antigenic map
}
