% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_agsr_matching.R
\name{matchStrains}
\alias{matchStrains}
\alias{match_mapAntigens}
\alias{match_mapSera}
\title{Find matching antigens or sera between 2 maps}
\usage{
match_mapAntigens(map1, map2, passage_matching = "ignore", warnings = TRUE)

match_mapSera(map1, map2, passage_matching = "ignore", warnings = TRUE)
}
\arguments{
\item{map1}{The names to match to.}

\item{map2}{The set of names from which to find matches.}
}
\value{
Returns the indices of matching strains in map 2, or NA in the position of strains not found.
}
\description{
Find matching antigens or sera between 2 maps
}
