% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_titers.R
\name{mergeTiterTables}
\alias{mergeTiterTables}
\title{Merge titer tables}
\usage{
mergeTiterTables(titer_tables)
}
\arguments{
\item{titer_tables}{A list of titer tables}
}
\value{
Returns a single merged titer table
}
\description{
Merges a list of titer tables into one titer table
}
\seealso{
Other {map merging functions}: 
\code{\link{mergeMaps}()},
\code{\link{mergeReport}()}
}
\concept{{map merging functions}}
