% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_merge.R
\name{merging_maps}
\alias{merging_maps}
\alias{mergeReport}
\alias{mergeMaps}
\title{Merging maps}
\usage{
mergeReport(map1, map2, passage_matching = "auto")

mergeMaps(map1, map2, method = "table", passage_matching = "auto",
  optimization_number_map1 = NULL, optimization_number_map2 = NULL,
  optimizations = 100)
}
\arguments{
\item{map1}{The first map}

\item{map2}{The second map}

\item{passage_matching}{Passage matching}

\item{method}{Merging method}

\item{optimization_number_map1}{Optimization number from the first map to merge}

\item{optimization_number_map2}{Optimization number from the second map to merge}

\item{optimizations}{Number of optimization runs to perform (for 'incremental method')}
}
\value{
Returns the merged map data or merge report.
}
\description{
Functions to merge together two tables or maps.
}
\details{
Maps can be merged in a number of ways depending upon the desired result.

\subsection{Method 'table'}{
This merges the tables of the two maps but does not attempt to create any new optimizations.
}

\subsection{Method 'incremental'}{
The default or specified optimization of the first map is copied to the resulting
map, optimizations of the second (merged in) map are ignored. Positions of
points that are in the merged-in map and not in the first map are
randomized and resulting layout relaxed.
}

\subsection{Method 'frozen'}{
This takes the default or specified optimization from the second map and
realigns it to the default or specifed optimization of the first map using
procrustes, then points that are found just in the second map receive
coordinates from that realigned optimization. Points that are common for the
first and the second map placed to their middle positions between the first
map layout and realigned second map layout.
}

\subsection{Method 'overlay'}{
Frozen merge followed by relaxation of the resulting optimization.
}
}
