% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimizations.R
\name{minColBasis}
\alias{minColBasis}
\title{Get the minimum column basis for an acmap}
\usage{
minColBasis(map, optimization_number = NULL)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization data to access (defaults to the currently selected optimization)}
}
\value{
Returns either the minimum column basis used or "fixed" if fixed column bases were specified.
}
\description{
Get the minimum column basis associated with the currently selected, or specified acmap optimization.
Once a optimization has been created the column bases cannot be further modified but instead a new
optimization must be created (see \code{\link{addOptimization}}).
}
\seealso{
\code{\link{colBases}} For getting the actual numeric column bases that have been assumed.

Other functions to get and set acmap optimization attributes: \code{\link{acmapCoords}},
  \code{\link{mapComments}}, \code{\link{mapDimensions}},
  \code{\link{mapStress}}, \code{\link{mapTransformation}}
}
\concept{functions to get and set acmap optimization attributes}
