% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{optimizationProperties}
\alias{optimizationProperties}
\alias{allMapStresses}
\alias{allMapDimensions}
\title{Get optimization properties}
\usage{
allMapStresses(map)

allMapDimensions(map)
}
\arguments{
\item{map}{map data object}
}
\description{
Utility functions to get a vector of all the map optimization
properties.
}
\seealso{
Other {functions to work with map optimizations}: 
\code{\link{keepSingleOptimization}()},
\code{\link{removeOptimizations}()},
\code{\link{sortOptimizations}()}
}
\concept{{functions to work with map optimizations}}
