% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimize.R
\name{optimizeSubsetAgsIndividually}
\alias{optimizeSubsetAgsIndividually}
\title{Optimize a subset of antigens individually}
\usage{
optimizeSubsetAgsIndividually(map, subset, ...)
}
\arguments{
\item{map}{The acmap object}

\item{subset}{A list of antigen names to optimize individually.}

\item{...}{Arguments passed to optimizeMap. number_of_dimensions and
number_of_optimizations are mandatory.}
}
\value{
List containing 'merge' and 'individual'. 'merge' is the reference
map with with the positions of antigens in the subset overlain.
'individual' is a list containing each of the individual maps.
}
\description{
You may have a large subset of antigens that distort the positions of
reference antigens and antisera. It can be useful to make a map containing
only reference antigens and antisera, and add each antigen in the subset to
the reference map individually. The resulting reference maps containing one
additional antigen are then aligned, and the positions of the additional
antigen in their individual map are shown.
}
\details{
A common use case might be if you want to introduce many tens, or even
hundreds, of mutants into an existing map, without the geometric patterns in
the mutant's titrations swamping those between the existing antigens and
sera.
}
\seealso{
Other {map optimization functions}: 
\code{\link{RacOptimizer.options}()},
\code{\link{make.acmap}()},
\code{\link{moveTrappedPoints}()},
\code{\link{optimizeMap}()},
\code{\link{randomizeCoords}()},
\code{\link{relaxMapOneStep}()},
\code{\link{relaxMap}()}
}
\concept{{map optimization functions}}
