% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_subset.R
\name{orderPoints}
\alias{orderPoints}
\alias{orderAntigens}
\alias{orderSera}
\title{Order antigens and sera}
\usage{
orderAntigens(map, order)

orderSera(map, order)
}
\arguments{
\item{map}{The map data object}

\item{order}{The new order of points}
}
\value{
An acmap object with points reordered
}
\description{
Functions to change the order of antigens and sera in a map
}
\seealso{
Other functions for working with map data:
\code{\link[=acmap]{acmap()}},
\code{\link[=addOptimization]{addOptimization()}},
\code{\link[=agReactivityAdjustments]{agReactivityAdjustments()}},
\code{\link[=as.json]{as.json()}},
\code{\link[=edit_agNames]{edit_agNames()}},
\code{\link[=edit_srNames]{edit_srNames()}},
\code{\link[=keepBestOptimization]{keepBestOptimization()}},
\code{\link[=keepSingleOptimization]{keepSingleOptimization()}},
\code{\link[=layerNames]{layerNames()}},
\code{\link[=read.acmap]{read.acmap()}},
\code{\link[=read.titerTable]{read.titerTable()}},
\code{\link{removePoints}},
\code{\link[=save.acmap]{save.acmap()}},
\code{\link[=save.coords]{save.coords()}},
\code{\link[=save.titerTable]{save.titerTable()}},
\code{\link{subsetCommonPoints}},
\code{\link[=subsetMap]{subsetMap()}}
}
\concept{functions for working with map data}
