% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.R
\name{plot.acmap}
\alias{plot.acmap}
\title{Plot an antigenic map}
\usage{
\method{plot}{acmap}(
  x,
  optimization_number = 1,
  xlim = NULL,
  ylim = NULL,
  plot_ags = TRUE,
  plot_sr = TRUE,
  plot_labels = FALSE,
  plot_blobs = TRUE,
  grid.col = "grey90",
  grid.margin.col = grid.col,
  fill.alpha = 0.8,
  outline.alpha = 0.8,
  padding = 1,
  cex = 1,
  ...
)
}
\arguments{
\item{optimization_number}{The optimization number to plot}

\item{xlim}{optional x axis limits}

\item{ylim}{optional y axis limits}

\item{plot_ags}{logical, should antigens be plotted}

\item{plot_sr}{logical, should antigens be plotted}

\item{plot_labels}{logical, should point labels be plotted}

\item{plot_blobs}{logical, should stress blobs be plotted if present}

\item{grid.col}{grid line color}

\item{grid.margin.col}{grid margin color}

\item{fill.alpha}{alpha for point fill}

\item{outline.alpha}{alpha for point outline}

\item{padding}{padding at limits of the antigenic map, ignored if xlim or ylim set explicitly}

\item{cex}{point size expansion factor}

\item{...}{additional arguments, not used}

\item{map}{The acmap to plot}
}
\description{
Method for plotting an antigenic map in two dimensions
}
\seealso{
Other {functions to view maps}: 
\code{\link{grid.plot.acmap}()},
\code{\link{snapshotMap}()}
}
\concept{{functions to view maps}}
