% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.R
\name{plot.rac}
\alias{plot.rac}
\title{Plot an antigenic map}
\usage{
\method{plot}{rac}(
  map,
  optimization_number = NULL,
  xlim = NULL,
  ylim = NULL,
  asp = 1,
  plot_ags = TRUE,
  plot_sr = TRUE,
  plot_labels = FALSE,
  grid.col = "grey90",
  grid.margin.col = grid.col,
  fill.alpha = 0.8,
  outline.alpha = 0.8,
  padding = 1,
  cex = 1
)
}
\description{
Method for plotting an antigenic map
}
\seealso{
Other {functions to view maps}: 
\code{\link{snapshotMap}()}
}
\concept{{functions to view maps}}
