% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_map_table_distance}
\alias{plot_map_table_distance}
\title{Plot map vs table distances}
\usage{
plot_map_table_distance(map, optimization_number = NULL, xlim, ylim)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization number from which to take map and table distances (defaults to the currently selected optimization)}

\item{xlim}{The x limits of the plot}

\item{ylim}{The y limits of the plot}
}
\value{
Silently returns information on the map and table distances
}
\description{
Plot map vs table distances
}
