% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_sr_titers}
\alias{plot_sr_titers}
\alias{plotly_sr_titers}
\title{Plot serum titers}
\usage{
plot_sr_titers(
  map,
  serum,
  xlim = NULL,
  ylim = NULL,
  optimization_number = NULL,
  .plot = TRUE
)

plotly_sr_titers(
  map,
  serum,
  xlim = NULL,
  ylim = NULL,
  optimization_number = NULL
)
}
\description{
Plot serum titers
}
\seealso{
Other Map Diagnostics: 
\code{\link{plot_stress_per_titer}}
}
\concept{Map Diagnostics}
