% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_stress_per_titer}
\alias{plot_stress_per_titer}
\alias{agMeanResiduals}
\title{Plot stress per titer}
\usage{
agMeanResiduals(map, exclude_nd = TRUE)
}
\description{
Plot stress per titer
}
\seealso{
Other Map Diagnostics: 
\code{\link{plot_sr_titers}()}
}
\concept{Map Diagnostics}
