% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_methods.R
\name{print.rac}
\alias{print.rac}
\title{Printing racmap objects}
\usage{
\method{print}{rac}(map, all = FALSE)
}
\arguments{
\item{racmap}{The map object}
}
\description{
Print information about a racmap object.
}
\examples{
# Print subset of information about a map
print(h3map2004)

# Print all information about a map
print(h3map2004, all = TRUE)

}
