% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_procrustes.R
\name{procrustesData}
\alias{procrustesData}
\title{Not yet implemented}
\usage{
procrustesData(
  map,
  comparison_map,
  optimization_number = 1,
  comparison_optimization_number = 1,
  translation = TRUE,
  scaling = FALSE
)
}
\description{
Not yet implemented
}
