% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_antigens.R
\name{ptAnnotations}
\alias{ptAnnotations}
\alias{agAnnotations}
\alias{srAnnotations}
\alias{agAnnotations<-}
\alias{srAnnotations<-}
\title{Getting and setting point annotation information}
\usage{
agAnnotations(map)

srAnnotations(map)

agAnnotations(map) <- value

srAnnotations(map) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{value}{A list of character vectors with annotations information for each
point}
}
\value{
A character vector of point annotations.
}
\description{
Getting and setting point annotation information
}
\seealso{
Other antigen and sera attribute functions:
\code{\link{agAttributes}},
\code{\link[=agGroups]{agGroups()}},
\code{\link[=agHomologousSr]{agHomologousSr()}},
\code{\link[=agLabIDs]{agLabIDs()}},
\code{\link[=agSequences]{agSequences()}},
\code{\link{ptClades}},
\code{\link{srAttributes}},
\code{\link[=srGroups]{srGroups()}},
\code{\link[=srHomologousAgs]{srHomologousAgs()}},
\code{\link[=srSequences]{srSequences()}}
}
\concept{antigen and sera attribute functions}
