% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_optimizations.R
\name{ptBaseCoords}
\alias{ptBaseCoords}
\alias{agBaseCoords}
\alias{srBaseCoords}
\alias{agBaseCoords<-}
\alias{srBaseCoords<-}
\title{Getting and setting base coordinates}
\usage{
agBaseCoords(map, optimization_number = NULL, .name = TRUE)
agBaseCoords(map, optimization_number = NULL, .check = TRUE) <- value

srBaseCoords(map, optimization_number = NULL, .name = TRUE)
srBaseCoords(map, optimization_number = NULL, .check = TRUE) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization run to get / set the data (by default the currently selected one)}

\item{.name}{Should the output be labelled with corresponding antigen / sera names}

\item{.check}{Should normal checks be applied for validity be applied when setting this value (used internally)}
}
\value{
Returns either the requested attribute when using a getter function or the updated acmap object when using the setter function.
}
\description{
These functions get and set the base coordinates for a given optimization run.
}
\seealso{
\code{\link{agCoords}}
\code{\link{srCoords}}

Other {map optimization attribute functions}: 
\code{\link{colBases}()},
\code{\link{mapComment}()},
\code{\link{mapDimensions}()},
\code{\link{mapStress}()},
\code{\link{mapTransformation}()},
\code{\link{ptCoords}()}
}
\concept{{map optimization attribute functions}}
