% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_blobs.R
\name{ptBootstrapBlob}
\alias{ptBootstrapBlob}
\alias{agBootstrapBlob}
\alias{srBootstrapBlob}
\alias{agBootstrapBlobs}
\alias{srBootstrapBlobs}
\alias{ptBootstrapBlobs}
\title{Get antigen or serum bootstrap blob information}
\usage{
agBootstrapBlob(map, antigen, optimization_number = 1)

srBootstrapBlob(map, serum, optimization_number = 1)

agBootstrapBlobs(map, optimization_number = 1)

srBootstrapBlobs(map, optimization_number = 1)

ptBootstrapBlobs(map, optimization_number = 1)
}
\arguments{
\item{map}{An acmap object}

\item{antigen}{The antigen to get the blob for}

\item{optimization_number}{Optimization number from which to get blob information}

\item{serum}{The serum to get the blob for}
}
\value{
Returns an object of class "blob" that can be plotted using the \code{blob()} function.
}
\description{
Get antigen or serum bootstrap blob information for plotting with the \code{blob()} function.
}
\seealso{
Other map diagnostic functions:
\code{\link[=agCohesion]{agCohesion()}},
\code{\link[=bootstrapBlobs]{bootstrapBlobs()}},
\code{\link[=bootstrapMap]{bootstrapMap()}},
\code{\link[=checkHemisphering]{checkHemisphering()}},
\code{\link[=dimensionTestMap]{dimensionTestMap()}},
\code{\link[=logtiterTable]{logtiterTable()}},
\code{\link{map-table-distances}},
\code{\link{mapBootstrapCoords}},
\code{\link[=mapDistances]{mapDistances()}},
\code{\link[=mapRelaxed]{mapRelaxed()}},
\code{\link[=mapResiduals]{mapResiduals()}},
\code{\link{pointStress}},
\code{\link[=ptBootstrapCoords]{ptBootstrapCoords()}},
\code{\link{ptLeverage}},
\code{\link{ptTriangulationBlob}},
\code{\link[=recalculateStress]{recalculateStress()}},
\code{\link[=stressTable]{stressTable()}},
\code{\link[=tableColbases]{tableColbases()}},
\code{\link[=tableDistances]{tableDistances()}},
\code{\link[=triangulationBlobs]{triangulationBlobs()}},
\code{\link{unstableMaps}}
}
\concept{map diagnostic functions}
