% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_diagnostics_bootstrap.R
\name{ptBootstrapCoords}
\alias{ptBootstrapCoords}
\alias{agBootstrapCoords}
\alias{srBootstrapCoords}
\title{Get antigen or serum bootstrap coordinates information}
\usage{
ptBootstrapCoords(map, point)

agBootstrapCoords(map, antigen)

srBootstrapCoords(map, serum)
}
\arguments{
\item{map}{An acmap object}

\item{point}{The point from which to get the bootstrap coords (numbered
antigens then sera)}

\item{antigen}{The antigen to get the bootstrap coords}

\item{serum}{The serum from which to get the bootstrap coords}
}
\value{
Returns a matrix of coordinates for the point in each of the
bootstrap runs
}
\description{
Get antigen or serum bootstrap coordinates information
}
\seealso{
Other map diagnostic functions:
\code{\link[=agCohesion]{agCohesion()}},
\code{\link[=bootstrapBlobs]{bootstrapBlobs()}},
\code{\link[=bootstrapMap]{bootstrapMap()}},
\code{\link[=checkHemisphering]{checkHemisphering()}},
\code{\link[=dimensionTestMap]{dimensionTestMap()}},
\code{\link[=logtiterTable]{logtiterTable()}},
\code{\link{map-table-distances}},
\code{\link{mapBootstrapCoords}},
\code{\link[=mapDistances]{mapDistances()}},
\code{\link[=mapRelaxed]{mapRelaxed()}},
\code{\link[=mapResiduals]{mapResiduals()}},
\code{\link{pointStress}},
\code{\link{ptBootstrapBlob}},
\code{\link{ptLeverage}},
\code{\link{ptTriangulationBlob}},
\code{\link[=recalculateStress]{recalculateStress()}},
\code{\link[=stressTable]{stressTable()}},
\code{\link[=tableColbases]{tableColbases()}},
\code{\link[=tableDistances]{tableDistances()}},
\code{\link[=triangulationBlobs]{triangulationBlobs()}},
\code{\link{unstableMaps}}
}
\concept{map diagnostic functions}
