% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_antigens.R
\name{ptClades}
\alias{ptClades}
\alias{agClades}
\alias{srClades}
\alias{agClades<-}
\alias{srClades<-}
\title{Getting and setting point clade information}
\usage{
agClades(map)

srClades(map)

agClades(map) <- value

srClades(map) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{value}{A list of character vectors with clade information for each
point}
}
\value{
A character vector of clade information.
}
\description{
Getting and setting point clade information
}
\seealso{
Other antigen and sera attribute functions: 
\code{\link{agAttributes}},
\code{\link{agGroups}()},
\code{\link{agHomologousSr}()},
\code{\link{agLabIDs}()},
\code{\link{agSequences}()},
\code{\link{ptAnnotations}},
\code{\link{srAttributes}},
\code{\link{srGroups}()},
\code{\link{srHomologousAgs}()},
\code{\link{srSequences}()}
}
\concept{antigen and sera attribute functions}
