% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_plotspec.R
\name{ptDrawingOrder}
\alias{ptDrawingOrder}
\alias{ptDrawingOrder<-}
\title{Get and set point drawing order in map}
\usage{
ptDrawingOrder(map)

ptDrawingOrder(map) <- value
}
\arguments{
\item{map}{An acmap object}

\item{value}{The point drawing order}
}
\value{
A numeric vector of point drawing order information
}
\description{
Point drawing order is a vector of indices defining the order in
which points should be draw when plotting or viewing a map. Points
are indexed in the same order as antigens then followed by
sera.
}
\seealso{
Other map point style functions: 
\code{\link{applyPlotspec}()},
\code{\link{ptOpacity}},
\code{\link{ptStyles}}
}
\concept{map point style functions}
