% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_plotspec.R
\name{ptOpacity}
\alias{ptOpacity}
\alias{agOpacity<-}
\alias{srOpacity<-}
\title{Set point opacity in a map}
\usage{
agOpacity(map) <- value

srOpacity(map) <- value
}
\arguments{
\item{map}{An acmap object}

\item{value}{A vector of opacities}
}
\value{
A numeric vector of point opacities.
}
\description{
These are helper functions to quickly set the opacity of points in a map,
they set both the fill and outline color opacity by modifying the fill
and outline colors to include an alpha channel for opacity. If you need
more control, for example different opacities for the fill and outline
colors, you alter the fill and outline opacities yourself, for example
with the \code{grDevices::adjustcolor()} function.
}
\seealso{
Other map point style functions: 
\code{\link{applyPlotspec}()},
\code{\link{ptDrawingOrder}()},
\code{\link{ptStyles}}
}
\concept{map point style functions}
