% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_plotspec.R
\name{ptStyles}
\alias{ptStyles}
\alias{agShown}
\alias{agSize}
\alias{agFill}
\alias{agOutline}
\alias{agOutlineWidth}
\alias{agRotation}
\alias{agAspect}
\alias{agShape}
\alias{srShown}
\alias{srSize}
\alias{srFill}
\alias{srOutline}
\alias{srOutlineWidth}
\alias{srRotation}
\alias{srAspect}
\alias{srShape}
\title{Getting and setting point plotting styles}
\usage{
agShown(map)

agSize(map)

agFill(map)

agOutline(map)

agOutlineWidth(map)

agRotation(map)

agAspect(map)

agShape(map)

srShown(map)

srSize(map)

srFill(map)

srOutline(map)

srOutlineWidth(map)

srRotation(map)

srAspect(map)

srShape(map)
}
\arguments{
\item{map}{The acmap data object}
}
\value{
Returns either the requested attribute when using a getter function or the updated acmap object when using the setter function.
}
\description{
These functions get and set the styles to use for each point when plotting.
}
\seealso{
Other {map point style functions}: 
\code{\link{applyPlotspec}()}
}
\concept{{map point style functions}}
