% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_blobs.R
\name{ptTriangulationBlobsize}
\alias{ptTriangulationBlobsize}
\alias{agTriangulationBlobSize}
\alias{srTriangulationBlobSize}
\title{Fetch information on stress blob size}
\usage{
agTriangulationBlobSize(map, optimization_number = 1)

srTriangulationBlobSize(map, optimization_number = 1)
}
\arguments{
\item{map}{acmap with stress blob information added}

\item{optimization_number}{optimization number for which to calculate blob size}
}
\description{
Returns a vector of stress blob sizes for each point, helpful for
programatically finding the points with the most uncertainty.
}
