% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Racmacs.R
\docType{package}
\name{Racmacs}
\alias{Racmacs}
\title{Racmacs: A package for performing antigenic cartography}
\description{
The Racmacs package provides a toolkit for making antigenic maps from assay
data such as HI assays, as described in Smith et al. 2004 (1).
}
\details{
\subsection{The acmap data object}{
The fundamental unit of the Racmacs package is the \link{acmap}
object, short for Antigenic Cartography MAP. This object contains all the
information about an antigenic map. You can read in a new acmap object from a
file with the function \link{read.acmap} and create a new acmap object within an
R session using the \link{acmap} function.

Key information associated with each acmap object is summarized in the sections below.
}

\subsection{The titer table}{
Each \link{acmap} object is built upon a table of data of measurements the
reactivity of a set of different sera against a set of different antigens.
Typically these measurements are HI assay measurements, but it is also
possible to use other similar assays as input data.

For the table of data it is always assumed that sera form the \emph{columns} and
antigens form the \emph{rows}, as below. You can get and set titer data with
\link{titerTable}.
}

\subsection{Optimizations}{
Another key component of the \link{acmap} object is a list of optimizations.
While acmap objects only have one table of data, they can have many
optimizations or none at all.

Each optimization has the following main attributes (see the vignette on
optimizing antigenic maps for more details on minimum column bases and
stress):
\itemize{
\item \strong{Antigen coordinates}, the coordinates of each antigen in this optimization.
\item \strong{Sera coordinates}, the coordinates of each serum in this optimization.
\item \strong{Minimum column basis}, the minimum column basis assumed when calculating this optimization.
\item \strong{Stress}, the stress of this optimization.
\item \strong{Dimensions}, the number of dimensions of this optimization.
}

A map may only have one optimization associated with it, simply representing
the optimal position of points in the map after number of optimization runs.
However it may also have a number of optimizations, perhaps representing the
different solutions for best antigen and serum positions based on different
random starting conditions in the optimizer, or perhaps giving the map
optimization in different numbers of dimensions.

At any one time a map object has one of the optimizations \emph{selected} (by
default the first one). Any information you read from this map that comes
from a optimization (for example \link{agCoords} ) and any functions that you
perform that relate to a optimization (for example \link{relaxMap}), will be
performed on the selected optimization by default. You can get and set the
selected optimization with \link{selectOptimization}.
}

\subsection{Plotting styles}{
The final type of information that is contained in the acmap object is
information on point styles when plotting. By altering these attributes you
can change the appearance of the antigen and serum points in any maps
plotted, the main ones include:
\itemize{
\item Size
\item Shape
\item Fill color
\item Outline color
}

}

\subsection{List of key functions}{

\strong{Reading and writing acmaps}
\itemize{
\item \link{read.acmap}
\item \link{save.acmap}
\item \link{save.coords}
\item \link{save.titerTable}
}

\strong{Optimizing maps}
\itemize{
\item \link{optimizeMap}
\item \link{relaxMap}
}

\strong{Get and set acmap information}
\itemize{
\item \link{titerTable}
\item \link{numAntigens}
\item \link{numSera}
\item \link{numPoints}
\item \link{numOptimizations}
\item \link{selectedOptimization}
}

\strong{Get and set plotting information}
\itemize{
\item \link{agFill}
\item \link{srFill}
\item \link{agOutline}
\item \link{srOutline}
\item \link{agOutlineWidth}
\item \link{srOutlineWidth}
\item \link{agSize}
\item \link{srSize}
\item \link{agShape}
\item \link{srShape}
}

\strong{Get and set optimization information}
\itemize{
\item \link{addOptimization}
\item \link{agCoords}
\item \link{srCoords}
\item \link{colBases}
\item \link{minColBasis}
\item \link{mapStress}
\item \link{mapDimensions}
}

\strong{Procrustes and realignment}
\itemize{
\item \link{procrustesMap}
\item \link{realignMap}
\item \link{realignOptimizations}
}

\strong{Diagnostic tests}
\itemize{
\item \link{checkHemisphering}
\item \link{moveTrappedPoints}
\item \link{dimensionTestMap}
}

\strong{Merging maps}
\itemize{
\item \link{mergeMaps}
\item \link{mergeReport}
}

}
}
