% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_viewer.R
\name{RacViewer.options}
\alias{RacViewer.options}
\title{Set viewer options}
\usage{
RacViewer.options(
  point.opacity = NA,
  viewer.controls = "hidden",
  grid.display = "static",
  grid.col = "#e0e0e0",
  show.names = FALSE,
  show.errorlines = FALSE,
  show.connectionlines = FALSE,
  show.titers = FALSE
)
}
\arguments{
\item{point.opacity}{Default opacity for unselected points}

\item{viewer.controls}{Should viewer controls be shown or hidden by default?}

\item{grid.display}{For 3d maps, should the grid be fixed in the background
or enclose and rotate along with the map}

\item{grid.col}{Color to use for the grid shown behind the map}

\item{show.names}{Toggle name labels on}

\item{show.errorlines}{Toggle error lines on}

\item{show.connectionlines}{Toggle connection lines on}

\item{show.titers}{Toggle titer labels on}
}
\value{
Returns a named list of viewer options
}
\description{
This function facilitates setting racviewer options by returning a list of
option settings.
}
\seealso{
Other {functions to view maps}: 
\code{\link{RacViewer}()},
\code{\link{export_viewer}()},
\code{\link{grid.plot.acmap}()},
\code{\link{plot.acmap}()},
\code{\link{setLegend}()},
\code{\link{snapshotMap}()},
\code{\link{view.acmap}()},
\code{\link{view.default}()},
\code{\link{view}()}
}
\concept{{functions to view maps}}
