% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RacViewer.R
\name{RacViewerProxy}
\alias{RacViewerProxy}
\title{Create a proxy RacViewer object}
\usage{
RacViewerProxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{ID of RacViewer instance}

\item{session}{Session object of shiny application}
}
\value{
Returns a chart proxy
}
\description{
Create a proxy RacViewer object
}
