% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_procrustes.R
\name{realignMap}
\alias{realignMap}
\title{Realign map to match another}
\usage{
realignMap(map, target_map, translation = TRUE, scaling = FALSE)
}
\arguments{
\item{map}{The acmap to realign.}

\item{target_map}{The acmap to realign to.}

\item{antigens}{Antigens to include when calculating the closest realignment, specified either by name or index or FALSE for excluding all.}

\item{sera}{Sera to include when calculating the closest realignment, specified either by name or index or FALSE for excluding all.}

\item{optimization_number}{The optimization from the map to realign (default is the currently selected one)}

\item{target_optimization_number}{The optimization from the target map to realign to (default is the currently selected one)}

\item{passage_matching}{The type of passage matching to be performed.}

\item{alignTargetToMain}{Should the target map be realigned to match the main map (default is to
realign main map to match target map) - this is sometimes useful if you want to
match based on antigen and serum indices from the main map.}
}
\value{
Returns the map aligned to the target map (or the target map aligned to the main map if alignTargetToMain is TRUE)
}
\description{
Realigns the coordinates of a map to match a target map as closely as possible.
}
\seealso{
Other {functions to compare maps}: 
\code{\link{procrustesMap}()},
\code{\link{realignOptimizations}()}
}
\concept{{functions to compare maps}}
