% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_procrustes.R
\name{realignOptimizations}
\alias{realignOptimizations}
\title{Realigns optimizations in the map to the current optimization}
\usage{
realignOptimizations(map, antigens = TRUE, sera = TRUE)
}
\arguments{
\item{map}{The acmap data}

\item{antigens}{The antigens to include in the realignment, TRUE for all FALSE for none, or specified by name or index}

\item{sera}{The sera to include in the realignment, TRUE for all FALSE for none, or specified by name or index}
}
\value{
Returns the map with realigned optimizations
}
\description{
Realigns optimizations in the map to the current optimization
}
