% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stress.R
\name{recalculateStress}
\alias{recalculateStress}
\title{Recalculate the stress associated with an acmap optimization}
\usage{
recalculateStress(map, optimization_number = 1)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization number}
}
\value{
Returns the recalculated map stress for a given optimization
}
\description{
Recalculates the stress associated with the currently selected or
user-specified optimization.
}
\seealso{
See \code{pointStress()} for getting the stress of individual points.

Other map diagnostic functions:
\code{\link[=agCohesion]{agCohesion()}},
\code{\link[=bootstrapBlobs]{bootstrapBlobs()}},
\code{\link[=bootstrapMap]{bootstrapMap()}},
\code{\link[=checkHemisphering]{checkHemisphering()}},
\code{\link[=dimensionTestMap]{dimensionTestMap()}},
\code{\link[=logtiterTable]{logtiterTable()}},
\code{\link{map-table-distances}},
\code{\link{mapBootstrapCoords}},
\code{\link[=mapDistances]{mapDistances()}},
\code{\link[=mapRelaxed]{mapRelaxed()}},
\code{\link[=mapResiduals]{mapResiduals()}},
\code{\link{pointStress}},
\code{\link{ptBootstrapBlob}},
\code{\link[=ptBootstrapCoords]{ptBootstrapCoords()}},
\code{\link{ptLeverage}},
\code{\link{ptTriangulationBlob}},
\code{\link[=stressTable]{stressTable()}},
\code{\link[=tableColbases]{tableColbases()}},
\code{\link[=tableDistances]{tableDistances()}},
\code{\link[=triangulationBlobs]{triangulationBlobs()}},
\code{\link{unstableMaps}}

Other functions relating to map stress calculation:
\code{\link[=logtiterTable]{logtiterTable()}},
\code{\link[=mapDistances]{mapDistances()}},
\code{\link[=mapResiduals]{mapResiduals()}},
\code{\link{pointStress}},
\code{\link[=stressTable]{stressTable()}},
\code{\link[=tableColbases]{tableColbases()}},
\code{\link[=tableDistances]{tableDistances()}}
}
\concept{functions relating to map stress calculation}
\concept{map diagnostic functions}
