% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_transformation.R
\name{reflectMap}
\alias{reflectMap}
\title{Reflect a map}
\usage{
reflectMap(map, axis = "x", optimization_number = NULL)
}
\arguments{
\item{map}{The acmap object}

\item{axis}{Axis of reflection}

\item{optimization_number}{The optimization number (or NULL to apply to all
optimizations)}
}
\value{
An acmap object with reflection applied
}
\description{
Reflects map coordinates
}
\seealso{
Other functions relating to map transformation: 
\code{\link{applyMapTransform}()},
\code{\link{rotateMap}()},
\code{\link{translateMap}()}
}
\concept{functions relating to map transformation}
