% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimize.R
\name{relaxMap}
\alias{relaxMap}
\title{Relax a map}
\usage{
relaxMap(map, optimization_number = NULL)
}
\arguments{
\item{map}{The acmap object}

\item{optimization_number}{The optimization to relax (defaults to the currently selected optimization)}
}
\value{
Returns an acmap object with the newly relaxed point coordinates.
}
\description{
Optimize antigen and serum positions starting from their current coordinates
in the selected or specified optimization.
}
\seealso{
See \code{\link{optimizeMap}} for performing new optimization runs from random starting coordinates.

Other {map optimization functions}: 
\code{\link{make.acmap}()},
\code{\link{moveTrappedPoints}()},
\code{\link{optimizeMap}()},
\code{\link{randomizeCoords}()},
\code{\link{relaxMapOneStep}()}
}
\concept{{map optimization functions}}
