% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimize.R
\name{relaxMapOneStep}
\alias{relaxMapOneStep}
\title{Relax a map one step in the optimiser}
\usage{
relaxMapOneStep(map, optimization_number = NULL)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The map optimization number (defaults to the currently selected optimization)}
}
\value{
Returns an updated map object
}
\description{
Relax a map one step in the optimiser
}
\seealso{
Other {map optimization functions}: 
\code{\link{make.acmap}()},
\code{\link{moveTrappedPoints}()},
\code{\link{optimizeMap}()},
\code{\link{randomizeCoords}()},
\code{\link{relaxMap}()}
}
\concept{{map optimization functions}}
