% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_subset.R
\name{removePoints}
\alias{removePoints}
\alias{removeAntigens}
\alias{removeSera}
\title{Remove antigens and sera}
\usage{
removeAntigens(map, antigens)

removeSera(map, sera)
}
\arguments{
\item{map}{The map data object}

\item{antigens}{Antigens to remove (specified by name or index)}

\item{sera}{Sera to remove (specified by name or index)}
}
\value{
An acmap object with points removed
}
\description{
Functions to remove antigens and sera from a map
}
\seealso{
Other functions for working with map data:
\code{\link[=acmap]{acmap()}},
\code{\link[=addOptimization]{addOptimization()}},
\code{\link[=agReactivityAdjustments]{agReactivityAdjustments()}},
\code{\link[=as.json]{as.json()}},
\code{\link[=edit_agNames]{edit_agNames()}},
\code{\link[=edit_srNames]{edit_srNames()}},
\code{\link[=keepBestOptimization]{keepBestOptimization()}},
\code{\link[=keepSingleOptimization]{keepSingleOptimization()}},
\code{\link[=layerNames]{layerNames()}},
\code{\link{orderPoints}},
\code{\link[=read.acmap]{read.acmap()}},
\code{\link[=read.titerTable]{read.titerTable()}},
\code{\link[=save.acmap]{save.acmap()}},
\code{\link[=save.coords]{save.coords()}},
\code{\link[=save.titerTable]{save.titerTable()}},
\code{\link{subsetCommonPoints}},
\code{\link[=subsetMap]{subsetMap()}}
}
\concept{functions for working with map data}
