% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rotation.R
\name{rotate_coords_by_degrees}
\alias{rotate_coords_by_degrees}
\title{Rotate a set of coordinates}
\usage{
rotate_coords_by_degrees(coords, degrees, axis = NULL)
}
\description{
Rotate a set of coordinates
}
