% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rotation.R
\name{rotate_coords_by_radians}
\alias{rotate_coords_by_radians}
\title{Rotate a set of coordinates}
\usage{
rotate_coords_by_radians(coords, radians, axis = NULL)
}
\description{
Rotate a set of coordinates
}
