% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rotation.R
\name{rotation_matrix_3D}
\alias{rotation_matrix_3D}
\title{Generate a 3 dimensional rotation matrix}
\usage{
rotation_matrix_3D(radians, axis = "z")
}
\description{
Generate a 3 dimensional rotation matrix
}
