% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_diagnostics_dimtest.R
\name{runDimensionTestMap}
\alias{runDimensionTestMap}
\title{Run dimtest result}
\usage{
runDimensionTestMap(
  map,
  dimensions_to_test = 1:5,
  test_proportion = 0.1,
  minimum_column_basis = "none",
  fixed_column_bases = rep(NA, numSera(map)),
  number_of_optimizations = 1000,
  replicates_per_dimension = 100,
  options = list()
)
}
\description{
Run dimtest result
}
