% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI.R
\name{runGUI}
\alias{runGUI}
\title{Open the Racmacs GUI}
\usage{
runGUI()
}
\value{
Nothing returned, called only for the side effect of starting the viewer.
}
\description{
This function opens the Racmacs GUI in a new window
}
\seealso{
Other shiny app functions: 
\code{\link{RacViewer-shiny}},
\code{\link{view.acmap}()}
}
\concept{shiny app functions}
