% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_loadsave.R
\name{save.coords}
\alias{save.coords}
\title{Save acmap coordinate data to a file}
\usage{
save.coords(map, filename, antigens = TRUE, sera = TRUE)
}
\arguments{
\item{map}{The acmap data object.}

\item{filename}{Path to the file.}
}
\description{
Saves acmap coordinate data of all or specified antigens and sera to a .csv file.
}
\seealso{
Other functions to create and save acmap objects: \code{\link{acmap}},
  \code{\link{read.acmap.cpp}}, \code{\link{read.acmap}},
  \code{\link{save.acmap}}, \code{\link{save.titerTable}}
}
\concept{functions to create and save acmap objects}
