% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_transformation.R
\name{scaleMap}
\alias{scaleMap}
\title{Scale a map}
\usage{
scaleMap(map, scaling, optimization_number = NULL)
}
\arguments{
\item{map}{The acmap object}

\item{scaling}{Scaling to apply}

\item{optimization_number}{The optimization number (by default the selected one)}
}
\description{
Scales map coordinates
}
\seealso{
Other {functions relating to map transformation}: 
\code{\link{applyMapTransform}()},
\code{\link{reflectMap}()},
\code{\link{rotateMap}()},
\code{\link{transformMap}()},
\code{\link{translateMap}()}
}
\concept{{functions relating to map transformation}}
