% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_chart.R
\name{selectedOptimization}
\alias{selectedOptimization}
\alias{selectedOptimization<-}
\title{Get or set the selected optimization}
\usage{
selectedOptimization(map)

selectedOptimization(map) <- value
}
\arguments{
\item{map}{The acmap data object}
}
\value{
Returns the selected optimization number.
}
\description{
Functions to get and set the currently selected optimization from an acmap object (see details).
}
\details{
At any one time a map object has one of the optimizations \emph{selected}
(by default the first one). Any information you read from this map that
comes from a optimization (for example \code{\link{agCoords}}) and any functions that
you perform that relate to a optimization (for example \code{\link{relaxMap}}), will
be performed on the selected optimization by default.
}
\seealso{
Other {functions for working with map data}: 
\code{\link{acmap}()},
\code{\link{as.cpp}()},
\code{\link{as.json}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{removePoints}},
\code{\link{save.acmap}()},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{subsetMap}()}
}
\concept{{functions for working with map data}}
