% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_legend.R
\name{setLegend}
\alias{setLegend}
\title{Set acmap legend}
\usage{
setLegend(map, legend, fill, style.bottom = "8px", style.right = "8px")
}
\arguments{
\item{map}{The acmap object}

\item{legend}{A character vector of legend labels}

\item{fill}{The fill color to be used with the boxes that appear alongside
the legend labels}

\item{style.bottom}{"bottom" style of the div, specifying how far from the
bottom of the viewport the bottom of the legend is spaced.}

\item{style.right}{"right" style of the div, specifying how far from the
right of the viewport the bottom of the legend is spaced.}
}
\value{
Returns the updated acmap object
}
\description{
This sets the acmap legend used when viewing a map for example.
}
\seealso{
Other functions to view maps: 
\code{\link{RacViewer.options}()},
\code{\link{RacViewer}()},
\code{\link{export_viewer}()},
\code{\link{ggplot.acmap}()},
\code{\link{mapGadget}()},
\code{\link{plot.acmap}()},
\code{\link{view.acmap}()},
\code{\link{view.default}()},
\code{\link{view}()}
}
\concept{functions to view maps}
