% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RacViewer.R
\name{snapshotMap}
\alias{snapshotMap}
\title{Create a map snapshot}
\usage{
snapshotMap(map, width = 800, height = 800, filename = NULL, ...)
}
\arguments{
\item{map}{The map data file}

\item{width}{Snapshot width}

\item{height}{Snapshot height}

\item{filename}{File to save image to}

\item{...}{Further parameters to pass to view}
}
\description{
Create a map snapshot
}
\seealso{
Other {functions to view maps}: 
\code{\link{RacViewer.options}()},
\code{\link{RacViewer}()},
\code{\link{export_viewer}()},
\code{\link{ggplot.acmap}()},
\code{\link{mapGadget}()},
\code{\link{plot.acmap}()},
\code{\link{setLegend}()},
\code{\link{view.default}()},
\code{\link{view}()}
}
\concept{{functions to view maps}}
