% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_chart.R
\name{sortOptimizations}
\alias{sortOptimizations}
\title{Sort optimizations by stress}
\usage{
sortOptimizations(map)
}
\arguments{
\item{map}{map data object}
}
\description{
Sorts all the optimization runs for a given map object by stress
(lowest to highest). Note that this is done by default when running
\code{\link[=optimizeMap]{optimizeMap()}}.
}
\seealso{
Other {functions to work with map optimizations}: 
\code{\link{keepSingleOptimization}()},
\code{\link{optimizationProperties}},
\code{\link{removeOptimizations}()}
}
\concept{{functions to work with map optimizations}}
