% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_merge.R
\name{splitTiterLayers}
\alias{splitTiterLayers}
\title{Split a map made up from titer layers into a list of separate maps each with a titer table
corresponding to one of the layers}
\usage{
splitTiterLayers(map)
}
\arguments{
\item{map}{An acmap object with titer table layers}
}
\description{
Split a map made up from titer layers into a list of separate maps each with a titer table
corresponding to one of the layers
}
\seealso{
Other map merging functions: 
\code{\link{RacMerge.options}()},
\code{\link{htmlMergeReport}()},
\code{\link{mergeMaps}()},
\code{\link{mergeReport}()}
}
\concept{map merging functions}
