% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_sera.R
\name{srAttributes}
\alias{srAttributes}
\alias{srIDs}
\title{Getting and setting sera attributes}
\usage{
srIDs(map)
}
\description{
These functions get and set the sera attributes for a map.
}
\seealso{
\code{\link{agAttributes}}

Other {antigen and sera attribute functions}: 
\code{\link{agAttributes}},
\code{\link{agGroups}()},
\code{\link{srGroups}()}
}
\concept{{antigen and sera attribute functions}}
