% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_sera.R
\name{srGroups}
\alias{srGroups}
\alias{srGroups<-}
\title{Getting and setting sera groups}
\usage{
srGroups(map)

srGroups(map) <- value
}
\description{
These functions get and set the sera groupings for a map.
}
\seealso{
Other {antigen and sera attribute functions}: 
\code{\link{agAttributes}},
\code{\link{agGroups}()},
\code{\link{srAttributes}}
}
\concept{{antigen and sera attribute functions}}
