% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_sera.R
\name{srSequences}
\alias{srSequences}
\alias{srSequences<-}
\alias{srNucleotideSequences}
\alias{srNucleotideSequences<-}
\title{Getting and setting sera sequence information}
\usage{
srSequences(map, missing_value = ".")

srSequences(map) <- value

srNucleotideSequences(map, missing_value = ".")

srNucleotideSequences(map) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{missing_value}{Character to use to fill in portions of the sequence matrix
where sequence data is missing.}

\item{value}{A character matrix of sequences with rows equal to the number of
sera}
}
\value{
A character matrix of sequences with rows equal to the number of
sera.
}
\description{
Getting and setting sera sequence information
}
\seealso{
Other antigen and sera attribute functions:
\code{\link{agAttributes}},
\code{\link[=agGroups]{agGroups()}},
\code{\link[=agHomologousSr]{agHomologousSr()}},
\code{\link[=agLabIDs]{agLabIDs()}},
\code{\link[=agSequences]{agSequences()}},
\code{\link{ptAnnotations}},
\code{\link{ptClades}},
\code{\link{srAttributes}},
\code{\link[=srGroups]{srGroups()}},
\code{\link[=srHomologousAgs]{srHomologousAgs()}}
}
\concept{antigen and sera attribute functions}
