% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_blobs.R
\name{stressBlobs}
\alias{stressBlobs}
\title{Add stress blob data to a map object}
\usage{
stressBlobs(
  map,
  stress_lim = 1,
  grid_spacing = 0.25,
  antigens = TRUE,
  sera = TRUE,
  .check_relaxation = TRUE
)
}
\arguments{
\item{map}{The acmap data object}

\item{stress_lim}{The blob stress limit}

\item{grid_spacing}{Grid spacing to use when calculating the blob}

\item{antigens}{Antigens to calculate blobs for (TRUE for all FALSE for none, or specified by name or index)}

\item{sera}{Sera to calculate blobs for (TRUE for all FALSE for none, or specified by name or index)}

\item{data}{The stress data to add (calculated if not provided)}

\item{optimization_number}{The optimization number (defaults to the currently selected one)}

\item{progress_fn}{Function to use for progress reporting}
}
\value{
Returns the acmap data object with stress blob information added
}
\description{
Add stress blob data to a map object
}
