% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_subset.R
\name{subsetCommonPoints}
\alias{subsetCommonPoints}
\alias{subsetCommonAgs}
\alias{subsetCommonSrGroups}
\title{Remove antigens and sera}
\usage{
subsetCommonAgs(maps)

subsetCommonSrGroups(maps)
}
\arguments{
\item{maps}{A list of map data objects}
}
\description{
Functions to subset a list of maps to include only antigens, antigen groups, sera
or serum groups that are in common between them.
}
\seealso{
Other functions for working with map data:
\code{\link[=acmap]{acmap()}},
\code{\link[=addOptimization]{addOptimization()}},
\code{\link[=agReactivityAdjustments]{agReactivityAdjustments()}},
\code{\link[=as.json]{as.json()}},
\code{\link[=edit_agNames]{edit_agNames()}},
\code{\link[=edit_srNames]{edit_srNames()}},
\code{\link[=keepBestOptimization]{keepBestOptimization()}},
\code{\link[=keepSingleOptimization]{keepSingleOptimization()}},
\code{\link[=layerNames]{layerNames()}},
\code{\link{orderPoints}},
\code{\link[=read.acmap]{read.acmap()}},
\code{\link[=read.titerTable]{read.titerTable()}},
\code{\link{removePoints}},
\code{\link[=save.acmap]{save.acmap()}},
\code{\link[=save.coords]{save.coords()}},
\code{\link[=save.titerTable]{save.titerTable()}},
\code{\link[=subsetMap]{subsetMap()}}
}
\concept{functions for working with map data}
