% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_subset.R
\name{subsetMap}
\alias{subsetMap}
\title{Subset an antigenic map}
\usage{
subsetMap(map, antigens = TRUE, sera = TRUE)
}
\arguments{
\item{map}{The antigenic map object}

\item{antigens}{Antigens to keep, defaults to all.}

\item{sera}{Sera to keep, defaults to all.}
}
\value{
Returns a new antigenic map containing only match antigens and sera
}
\description{
Subset an antigenic map to contain only specified antigens and sera
}
\seealso{
Other functions for working with map data:
\code{\link[=acmap]{acmap()}},
\code{\link[=addOptimization]{addOptimization()}},
\code{\link[=agReactivityAdjustments]{agReactivityAdjustments()}},
\code{\link[=as.json]{as.json()}},
\code{\link[=edit_agNames]{edit_agNames()}},
\code{\link[=edit_srNames]{edit_srNames()}},
\code{\link[=keepBestOptimization]{keepBestOptimization()}},
\code{\link[=keepSingleOptimization]{keepSingleOptimization()}},
\code{\link[=layerNames]{layerNames()}},
\code{\link{orderPoints}},
\code{\link[=read.acmap]{read.acmap()}},
\code{\link[=read.titerTable]{read.titerTable()}},
\code{\link{removePoints}},
\code{\link[=save.acmap]{save.acmap()}},
\code{\link[=save.coords]{save.coords()}},
\code{\link[=save.titerTable]{save.titerTable()}},
\code{\link{subsetCommonPoints}}
}
\concept{functions for working with map data}
