% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimize.R
\name{subsetMapToConnectedGroup}
\alias{subsetMapToConnectedGroup}
\title{Subset the map to a single connected group}
\usage{
subsetMapToConnectedGroup(map, group_number = 1)
}
\arguments{
\item{map}{An antigenic map object}

\item{group_number}{The number of the connected group to subset to. Connected'
groups are ordered by size, so group 1 is the largest connected group of points
in the map.}
}
\value{
A subsetted map object containing only the antigens and sera in
the specified connected group.
}
\description{
Sometimes a map may have multiple different connected groups of points.
Where some groups of antigens and sera are not connected to each other by
any path of measurable titers, which will be default throw an error when
you try and optimize the map. For maps of this type, this function can be
used to subset the map to a single connected group of antigens and sera.
}
