% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_diagnostics_dimtest.R
\name{summary.ac_dimtest}
\alias{summary.ac_dimtest}
\title{Summarize dimension test results}
\usage{
\method{summary}{ac_dimtest}(object, ...)
}
\arguments{
\item{object}{An object as returned by \code{dimensionTestMap()}}

\item{...}{Additional unused arguments}
}
\value{
Returns a data frame with the following columns. "dimensions" : the
dimension tested, "mean_rmse_detectable" : mean prediction rmse for
detectable titers across all runs. "var_rmse_detectable" the variance of
the prediction rmse for detectable titers across all runs, useful for
estimating confidence intervals. "mean_rmse_nondetectable" and
"var_rmse_nondetectable" the equivalent for non-detectable titers
}
\description{
S3 method for class 'ac_dimtest', returning information on the
cross-validation results for each dimension tested.
}
\details{
For each run, the ag-sr titers that were randomly excluded are
predicted according to their relative positions in the map trained without
them. An RMSE is then calculated by comparing predicted titers inferred
from the map on the log scale to the actual log titers. This is done
separately for detectable titers (e.g. 40) and non-detectable titers (e.g.
<10). For non-detectable titers, if the predicted titer is the same or
lower than the log-titer threshold, the error is set to 0.
}
