% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{surface_tri_norm}
\alias{surface_tri_norm}
\title{Function to get surface triangles.}
\usage{
surface_tri_norm(coords, mesh_indices, check_fn, ...)
}
\arguments{
\item{coords}{Coordinates}

\item{mesh_indices}{Mesh indices}

\item{check_fn}{Checking function}

\item{...}{Additional parameters}
}
\description{
This uses the surf.tri function but ensures that all triangles are outward facing.
}
