% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_chart.R
\name{titerTableLayers}
\alias{titerTableLayers}
\title{Getting and setting titer table layers}
\usage{
titerTableLayers(map, .name = TRUE)
titerTableLayers(map, .check = TRUE) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{.name}{Should the output be labelled with corresponding antigen / sera names}

\item{.check}{Should normal checks be applied for validity be applied when setting this value (used internally)}
}
\value{
Returns either the requested attribute when using a getter function or the updated acmap object when using the setter function.
}
\description{
Functions to get and set the underlying titer table layers of a map (see details).
}
\seealso{
Other {map attribute functions}: 
\code{\link{name}},
\code{\link{titerTable}()}
}
\concept{{map attribute functions}}
