% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{titertypesTableLayers}
\alias{titertypesTableLayers}
\title{Return a list of titertypes tables}
\usage{
titertypesTableLayers(map)
}
\arguments{
\item{map}{An acmap data object}
}
\description{
Return a list of titertypes tables
}
