% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_transformation.R
\name{transformMap}
\alias{transformMap}
\title{Transform a map}
\usage{
transformMap(map, transformation_matrix, optimization_number = NULL)
}
\arguments{
\item{map}{The acmap object}

\item{transformation_matrix}{Transformation to apply (as matrix)}

\item{optimization_number}{The optimization number (by default the selected one)}
}
\description{
Transforms map coordinates
}
\seealso{
Other {functions relating to map transformation}: 
\code{\link{applyMapTransform}()},
\code{\link{reflectMap}()},
\code{\link{rotateMap}()},
\code{\link{scaleMap}()},
\code{\link{translateMap}()}
}
\concept{{functions relating to map transformation}}
