% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RacViewer.R
\name{unbox_map}
\alias{unbox_map}
\title{Converting map data to the json format
Function to define what parts of the map data should be unboxed when exporting to json format.}
\usage{
unbox_map(mapData)
}
\description{
Converting map data to the json format
Function to define what parts of the map data should be unboxed when exporting to json format.
}
