% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_edit_names.R
\name{update_ferret_seraNames}
\alias{update_ferret_seraNames}
\title{Update the ferret serum names to match antigens}
\usage{
update_ferret_seraNames(map, dictionary_file = NULL)
}
\arguments{
\item{map}{The map data object}

\item{dictionary_file}{The path to the dictionary file you want to use
(should be .csv). If not supplied then the default dictionary that comes
with Racmacs will be used.}
}
\value{
Returns the updated map data object
}
\description{
Update the ferret serum names to match antigens
}
\keyword{internal}
