% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_methods.R
\name{view.default}
\alias{view.default}
\title{Default method for viewing objects}
\usage{
\method{view}{default}(x, ...)
}
\arguments{
\item{x}{The object to view}

\item{...}{Additional arguments, passed to print.}
}
\value{
No value returned, simply calls the print method on the object
}
\description{
Default method for viewing objects
}
\seealso{
Other functions to view maps:
\code{\link[=RacViewer]{RacViewer()}},
\code{\link[=RacViewer.options]{RacViewer.options()}},
\code{\link[=export_viewer]{export_viewer()}},
\code{\link[=ggplot.acmap]{ggplot.acmap()}},
\code{\link[=mapGadget]{mapGadget()}},
\code{\link[=plot.acmap]{plot.acmap()}},
\code{\link[=setLegend]{setLegend()}},
\code{\link[=view]{view()}},
\code{\link[=view.acmap]{view.acmap()}}
}
\concept{functions to view maps}
