% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_methods.R
\name{view.rac}
\alias{view.rac}
\title{Viewing racmap objects}
\usage{
\method{view}{rac}(map, ..., .jsCode = NULL, .jsData = NULL,
  select_ags = NULL, select_sr = NULL, show_procrustes = NULL)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{view_map}}}

\item{racmap}{The racmap object}
}
\description{
View a racmap object in the interactive viewer.
}
