% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_blobs.R
\name{view.racblobs}
\alias{view.racblobs}
\title{Viewing map stress blob data}
\usage{
\method{view}{racblobs}(object, ...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{view.rac}}}

\item{map}{The map after stressBlobs() has been applied}
}
\description{
View stress blob data in an interactive viewer.
}
