% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_procrustes.R
\name{view.racprocrustes}
\alias{view.racprocrustes}
\title{Viewing procrustes data}
\usage{
\method{view}{racprocrustes}(object, ...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{view_map}}}

\item{map}{The primary map}

\item{target_map}{The target map}
}
\description{
View procrustes data in an interactive viewer.
}
