% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_methods.R
\name{view}
\alias{view}
\title{S3 method for viewing objects}
\usage{
view(x, ...)
}
\arguments{
\item{x}{The object to view}

\item{...}{Additional arguments, not used.}
}
\value{
When called on an acmap object, returns an htmlwidget object that
can be used to interactively view the map. Otherwise by default it
simply calls the print method of the respective object with no return
value.
}
\description{
S3 method for viewing objects
}
\seealso{
Other functions to view maps:
\code{\link[=RacViewer]{RacViewer()}},
\code{\link[=RacViewer.options]{RacViewer.options()}},
\code{\link[=export_viewer]{export_viewer()}},
\code{\link[=ggplot.acmap]{ggplot.acmap()}},
\code{\link[=mapGadget]{mapGadget()}},
\code{\link[=plot.acmap]{plot.acmap()}},
\code{\link[=setLegend]{setLegend()}},
\code{\link[=view.acmap]{view.acmap()}},
\code{\link[=view.default]{view.default()}}
}
\concept{functions to view maps}
