% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_viewer.R
\name{write2viewer_debug}
\alias{write2viewer_debug}
\title{Write map data to the viewer debug file}
\usage{
write2viewer_debug(map, snapshot = FALSE)
}
\description{
Write map data to the viewer debug file
}
