% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{acmapAttributes}
\alias{acmapAttributes}
\alias{numAntigens}
\alias{numSera}
\alias{numPoints}
\alias{numOptimizations}
\title{Get acmap attributes}
\usage{
numAntigens(map)

numSera(map)

numPoints(map)

numOptimizations(map)
}
\arguments{
\item{map}{The acmap data object}
}
\description{
Functions to get various attributes about an acmap object.
}
\seealso{
Other {map attribute functions}: 
\code{\link{adjustedLogTiterTable}()},
\code{\link{adjustedTiterTable}()},
\code{\link{dilutionStepsize}()},
\code{\link{mapName}()},
\code{\link{titerTableLayers}()},
\code{\link{titerTable}()}
}
\concept{{map attribute functions}}
