% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_chart.R
\name{acmapAttributes}
\alias{acmapAttributes}
\alias{numAntigens}
\alias{numSera}
\alias{numPoints}
\alias{numOptimizations}
\title{Get acmap attributes}
\usage{
numAntigens(map)

numSera(map)

numPoints(map)

numOptimizations(map)
}
\arguments{
\item{map}{The acmap data object}
}
\description{
Functions to get various attributes about an acmap object.
}
\concept{functions to get and set map attributes}
