% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{acmapAttributes}
\alias{acmapAttributes}
\alias{numAntigens}
\alias{numSera}
\alias{numSeraGroups}
\alias{numPoints}
\alias{numOptimizations}
\alias{numLayers}
\title{Get acmap attributes}
\usage{
numAntigens(map)

numSera(map)

numSeraGroups(map)

numPoints(map)

numOptimizations(map)

numLayers(map)
}
\arguments{
\item{map}{The acmap data object}
}
\value{
A number relating to the attribute
}
\description{
Functions to get various attributes about an acmap object.
}
\seealso{
Other map attribute functions: 
\code{\link{adjustedLogTiterTable}()},
\code{\link{adjustedTiterTable}()},
\code{\link{dilutionStepsize}()},
\code{\link{logtiterTableLayers}()},
\code{\link{mapDescription}()},
\code{\link{mapName}()},
\code{\link{titerTableLayers}()},
\code{\link{titerTable}()}
}
\concept{map attribute functions}
