% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reactivity_adjusts.R
\name{adjustedTiterTable}
\alias{adjustedTiterTable}
\title{Get the reactivity adjusted titer table}
\usage{
adjustedTiterTable(map, optimization_number = 1)
}
\arguments{
\item{map}{An acmap object}

\item{optimization_number}{The optimization number from which
to take any antigen reactivity adjustments}
}
\value{
A character matrix of titers.
}
\description{
Return the titer table plus any antigen reactivity adjustments.
}
\seealso{
\code{\link[=htmlAdjustedTiterTable]{htmlAdjustedTiterTable()}}

Other map attribute functions: 
\code{\link{acmapAttributes}},
\code{\link{adjustedLogTiterTable}()},
\code{\link{dilutionStepsize}()},
\code{\link{logtiterTableLayers}()},
\code{\link{mapDescription}()},
\code{\link{mapName}()},
\code{\link{titerTableFlat}()},
\code{\link{titerTableLayers}()},
\code{\link{titerTable}()}
}
\concept{map attribute functions}
