% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_antigens.R
\name{agAttributes}
\alias{agAttributes}
\alias{agIDs}
\alias{agNames}
\alias{agNames<-}
\alias{agIDs<-}
\alias{agNamesFull}
\alias{agNamesAbbreviated}
\alias{agDates}
\alias{agDates<-}
\alias{agReference}
\title{Getting and setting antigen attributes}
\usage{
agNames(map)

agNames<-(map)

agIDs(map)

agIDs<-(map)

agNamesFull(map)

agNamesAbbreviated(map)

agDates(map)

agDates<-(map)

agReference(map)
}
\arguments{
\item{map}{The acmap data object}
}
\value{
Returns either the requested attribute when using a getter function or the updated acmap object when using the setter function.
}
\description{
These functions get and set the antigen attributes for a map.
}
\seealso{
\code{\link{srAttributes}}

Other {antigen and sera attribute functions}: 
\code{\link{agGroups}()},
\code{\link{srAttributes}},
\code{\link{srGroups}()}
}
\concept{{antigen and sera attribute functions}}
