% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_antigens.R
\name{agAttributes}
\alias{agAttributes}
\alias{agIDs}
\alias{agNames}
\alias{agNames<-}
\alias{agIDs<-}
\alias{agDates}
\alias{agDates<-}
\alias{agNamesFull}
\alias{agNamesAbbreviated}
\alias{agReference}
\alias{agPassage}
\title{Getting and setting antigen attributes}
\usage{
agNames(map)
agNames(map) <- value
agIDs(map)
agIDs(map) <- value
agDates(map)
agDates(map) <- value
agNamesFull(map)
agNamesAbbreviated(map)
agReference(map)
agPassage(map)
}
\arguments{
\item{map}{The acmap data object}

\item{value}{New value to set}
}
\value{
Returns either the requested attribute when using a getter function or the updated acmap object

when using the setter function.
}
\description{
These functions get and set the antigen attributes for a map.
}
\seealso{
\code{srAttributes()}

Other {antigen and sera attribute functions}: 
\code{\link{agGroups}()},
\code{\link{agSequences}()},
\code{\link{ptClades}},
\code{\link{srAttributes}},
\code{\link{srGroups}()},
\code{\link{srSequences}()}
}
\concept{{antigen and sera attribute functions}}
