% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_antigens.R
\name{agAttributes}
\alias{agAttributes}
\alias{agIDs}
\alias{agIDs<-}
\alias{agDates}
\alias{agDates<-}
\alias{agReference}
\alias{agReference<-}
\alias{agNames}
\alias{agNames<-}
\alias{agExtra}
\alias{agExtra<-}
\alias{agPassage}
\alias{agPassage<-}
\alias{agLineage}
\alias{agLineage<-}
\alias{agReassortant}
\alias{agReassortant<-}
\alias{agStrings}
\alias{agStrings<-}
\alias{agContinent}
\alias{agContinent<-}
\title{Getting and setting antigen attributes}
\usage{
agIDs(map)
agIDs(map) <- value
agDates(map)
agDates(map) <- value
agReference(map)
agReference(map) <- value
agNames(map)
agNames(map) <- value
agExtra(map)
agExtra(map) <- value
agPassage(map)
agPassage(map) <- value
agLineage(map)
agLineage(map) <- value
agReassortant(map)
agReassortant(map) <- value
agStrings(map)
agStrings(map) <- value
agContinent(map)
agContinent(map) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{value}{New value to set}
}
\value{
Returns either the requested attribute when using a getter function or

the updated acmap object when using the setter function.
}
\description{
These functions get and set the antigen attributes for a map.
}
\seealso{
\code{srAttributes()}

Other antigen and sera attribute functions: 
\code{\link{agGroups}()},
\code{\link{agHomologousSr}()},
\code{\link{agLabIDs}()},
\code{\link{agSequences}()},
\code{\link{ptAnnotations}},
\code{\link{ptClades}},
\code{\link{srAttributes}},
\code{\link{srGroups}()},
\code{\link{srHomologousAgs}()},
\code{\link{srSequences}()}
}
\concept{antigen and sera attribute functions}
