% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_antigens.R
\name{agGroups}
\alias{agGroups}
\alias{agGroups<-}
\title{Getting and setting antigen groups}
\usage{
agGroups(map)

agGroups(map) <- value
}
\arguments{
\item{map}{The acmap object}

\item{value}{A character or factor vector of groupings to apply to the
antigens}
}
\value{
A factor vector of groupings.
}
\description{
These functions get and set the antigen groupings for a map.
}
\seealso{
Other antigen and sera attribute functions: 
\code{\link{agAttributes}},
\code{\link{agHomologousSr}()},
\code{\link{agLabIDs}()},
\code{\link{agSequences}()},
\code{\link{ptAnnotations}},
\code{\link{ptClades}},
\code{\link{srAttributes}},
\code{\link{srGroups}()},
\code{\link{srHomologousAgs}()},
\code{\link{srSequences}()}
}
\concept{antigen and sera attribute functions}
