% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{agReactivityAdjustments}
\alias{agReactivityAdjustments}
\alias{agReactivityAdjustments<-}
\title{Get and set antigen reactivity adjustments}
\usage{
agReactivityAdjustments(map)

agReactivityAdjustments(map) <- value
}
\arguments{
\item{map}{The acmap object}

\item{value}{A vector of antigen reactivity adjustments to apply}
}
\value{
A numeric vector of antigen reactivity adjustments
}
\description{
Get and set antigen reactivity adjustments
}
\seealso{
Other functions for working with map data: 
\code{\link{acmap}()},
\code{\link{addOptimization}()},
\code{\link{as.json}()},
\code{\link{edit_agNames}()},
\code{\link{edit_srNames}()},
\code{\link{keepBestOptimization}()},
\code{\link{keepSingleOptimization}()},
\code{\link{layerNames}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{removePoints}},
\code{\link{save.acmap}()},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{subsetCommonPoints}},
\code{\link{subsetMap}()}
}
\concept{functions for working with map data}
