% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_antigens.R
\name{agSequences}
\alias{agSequences}
\alias{agSequences<-}
\title{Getting and setting antigen sequence information}
\usage{
agSequences(map)

agSequences(map) <- value
}
\arguments{
\item{map}{The acmap data object}
}
\description{
Getting and setting antigen sequence information
}
