% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_antigens.R
\name{agSequences}
\alias{agSequences}
\alias{agSequences<-}
\alias{agNucleotideSequences}
\alias{agNucleotideSequences<-}
\title{Getting and setting antigen sequence information}
\usage{
agSequences(map, missing_value = ".")

agSequences(map) <- value

agNucleotideSequences(map, missing_value = ".")

agNucleotideSequences(map) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{missing_value}{Character to use to fill in portions of the sequence matrix
where sequence data is missing.}

\item{value}{A character matrix of sequences with rows equal to the number of
antigens}
}
\value{
A character matrix of sequences, where each row represents an antigen.
}
\description{
Getting and setting antigen sequence information
}
\seealso{
Other antigen and sera attribute functions: 
\code{\link{agAttributes}},
\code{\link{agGroups}()},
\code{\link{agHomologousSr}()},
\code{\link{agLabIDs}()},
\code{\link{ptAnnotations}},
\code{\link{ptClades}},
\code{\link{srAttributes}},
\code{\link{srGroups}()},
\code{\link{srHomologousAgs}()},
\code{\link{srSequences}()}
}
\concept{antigen and sera attribute functions}
