% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_conversion.R
\name{as.json}
\alias{as.json}
\title{Convert map to json format}
\usage{
as.json(map)
}
\arguments{
\item{map}{The map data object}
}
\value{
Returns map data as .ace json format
}
\description{
Convert map to json format
}
