% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_loadsave.R
\name{as.json}
\alias{as.json}
\title{Convert map to json format}
\usage{
as.json(map, pretty = FALSE, round_titers = FALSE)
}
\arguments{
\item{map}{The map data object}

\item{pretty}{Should json be output prettily with new lines and indentation?}

\item{round_titers}{Should titers be rounded to the nearest integer before outputting}
}
\value{
Returns map data as .ace json format
}
\description{
Convert map to json format
}
\seealso{
Other functions for working with map data: 
\code{\link{acmap}()},
\code{\link{addOptimization}()},
\code{\link{agReactivityAdjustments}()},
\code{\link{edit_agNames}()},
\code{\link{edit_srNames}()},
\code{\link{keepBestOptimization}()},
\code{\link{keepSingleOptimization}()},
\code{\link{layerNames}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{removePoints}},
\code{\link{save.acmap}()},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{subsetCommonPoints}},
\code{\link{subsetMap}()}
}
\concept{functions for working with map data}
