% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_bootstrap.R
\name{bootstrapMap}
\alias{bootstrapMap}
\title{Perform a noisy bootstrap}
\usage{
bootstrapMap(
  map,
  bootstrap_repeats = 1000,
  optimizations_per_repeat = 100,
  ag_noise_sd = 0.7,
  titer_noise_sd = 0.7,
  .progress = NULL
)
}
\arguments{
\item{map}{The map object}

\item{bootstrap_repeats}{The number of bootstrap repeats to perform}

\item{optimizations_per_repeat}{When reoptimizing the map from scratch, the number of optimization runs to perform}

\item{ag_noise_sd}{The standard deviation to use when applying noise per antigen}

\item{titer_noise_sd}{The standard deviation to use when applying noise per titer}
}
\value{
Returns the updated map object
}
\description{
Perform a noisy bootstrap on a map, adding random noise and reoptimizing from scratch.
}
\seealso{
Other {map diagnostic functions}: 
\code{\link{checkHemisphering}()},
\code{\link{dimensionTestMap}()},
\code{\link{mapRelaxed}()}
}
\concept{{map diagnostic functions}}
