% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_new.R
\name{cloneMap}
\alias{cloneMap}
\title{Clone an acmap object}
\usage{
cloneMap(map)
}
\arguments{
\item{map}{The map object}
}
\value{
Returns a copy of the map object
}
\description{
Creates a copy of an acmap object. This is needed, because acmap
objects, being environments rather than lists are not copy-on-modify, i.e.
if you change one reference to the acmap, all references to it will be changed.
To avoid this behaviour you can use the cloneMap function.
}
