% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_messages.R
\name{deprecated_functions}
\alias{deprecated_functions}
\alias{stressBlobs}
\title{Deprecated functions}
\usage{
stressBlobs(...)
}
\arguments{
\item{...}{Arguments to pass to the new function}
}
\value{
Values from the new function
}
\description{
These functions still work but have been deprecated in favour of another function. Arguments will be passed onto the new function with a warning.
}
