% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_diagnostics.R
\name{dimensionTestMap}
\alias{dimensionTestMap}
\title{Perform dimension testing on a map object}
\usage{
dimensionTestMap(map, dimensions_to_test = 1:5, test_proportion = 0.1,
  minimum_column_basis = "none", column_bases_from_master = TRUE)
}
\arguments{
\item{map}{The acmap data object}

\item{dimensions_to_test}{A numeric vector of dimensions to be tested}

\item{test_proportion}{The proportion of data to be used as the test set for each test run}

\item{minimum_column_basis}{The minimum column basis to use}

\item{column_bases_from_master}{Should column bases be calculated based on
the full data set? If false, they will be recalculated based on the
training set used for each run.}
}
\value{
Returns a table of results, including the dimensions tested and the standard deviation of predicted vs residual error.
}
\description{
Take a map object and perform cross-validation, seeing how well titers are
predicted when they are excluded from the map.
}
\concept{functions to test an acmap}
