% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_edit_names.R
\name{edit_agNames}
\alias{edit_agNames}
\title{Edit antigen names in an acmap}
\usage{
edit_agNames(map, old_names, new_names)
}
\arguments{
\item{map}{The map data object to be updated}

\item{old_names}{Old names to be replaced}

\item{new_names}{Replacement for old names}
}
\value{
Returns the acmap object with antigen names updated.
}
\description{
Edit antigen names in an acmap
}
\seealso{
Other functions for working with map data: 
\code{\link{acmap}()},
\code{\link{addOptimization}()},
\code{\link{agReactivityAdjustments}()},
\code{\link{as.json}()},
\code{\link{edit_srNames}()},
\code{\link{keepBestOptimization}()},
\code{\link{keepSingleOptimization}()},
\code{\link{layerNames}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{removePoints}},
\code{\link{save.acmap}()},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{subsetCommonPoints}},
\code{\link{subsetMap}()}
}
\concept{functions for working with map data}
