% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_edit_names.R
\name{edit_srNames}
\alias{edit_srNames}
\title{Edit sera names in a mapData object}
\usage{
edit_srNames(map, old_names, new_names, warnings = TRUE)
}
\arguments{
\item{map}{The map data object to be updated}

\item{old_names}{Old names to be replaced}

\item{new_names}{Replacement for old names}

\item{warnings}{Should a warning be issued if old names were not found in the map.}
}
\value{
Returns an updated map data object
}
\description{
Edit sera names in a mapData object
}
