% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_edit_names.R
\name{edit_srNames}
\alias{edit_srNames}
\title{Edit sera names in an acmap}
\usage{
edit_srNames(map, old_names, new_names, warnings = TRUE)
}
\arguments{
\item{map}{The map data object to be updated}

\item{old_names}{Old names to be replaced}

\item{new_names}{Replacement for old names}
}
\value{
Returns the acmap object with sera names updated.
}
\description{
Edit sera names in an acmap
}
