% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_viewer.R
\name{export_viewer}
\alias{export_viewer}
\title{Export the map viewer}
\usage{
export_viewer(map, file, selfcontained = TRUE, ...)
}
\arguments{
\item{map}{The acmap object}

\item{file}{html file to output to}

\item{selfcontained}{Self-contained html file}
}
\description{
Export a map in a standalone html viewer
}
