% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimizations.R
\name{getOptimization}
\alias{getOptimization}
\title{Get optimization details from an acmap object}
\usage{
getOptimization(map, optimization_number = 1)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization data to access}
}
\value{
Returns a list with information about the optimization
}
\description{
Gets the details associated with the currently selected or specified acmap
optimization as a list.
}
\seealso{
See \code{listOptimizations()} for getting information about all
optimizations.
}
