% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot_gridplot.R
\name{grid.plot.acmap}
\alias{grid.plot.acmap}
\title{Plot an antigenic map using the grid system}
\usage{
grid.plot.acmap(
  map,
  optimization_number = 1,
  xlim = NULL,
  ylim = NULL,
  plot_ags = TRUE,
  plot_sr = TRUE,
  plot_labels = FALSE,
  plot_blobs = TRUE,
  plot_hemisphering = TRUE,
  grid.col = "grey90",
  grid.margin.col = grid.col,
  fill.alpha = 0.8,
  outline.alpha = 0.8,
  padding = 1,
  cex = 1
)
}
\arguments{
\item{map}{The acmap to plot}

\item{optimization_number}{The optimization number to plot}

\item{xlim}{optional x axis limits}

\item{ylim}{optional y axis limits}

\item{plot_ags}{logical, should antigens be plotted}

\item{plot_sr}{logical, should antigens be plotted}

\item{plot_labels}{logical, should point labels be plotted}

\item{plot_blobs}{logical, should stress blobs be plotted if present}

\item{plot_hemisphering}{logical, should arrows be plotted to indicate hemisphering
or trapped points (only for maps where \code{checkHemisphering()} has already
been run)}

\item{grid.col}{grid line color}

\item{grid.margin.col}{grid margin color}

\item{fill.alpha}{alpha for point fill}

\item{outline.alpha}{alpha for point outline}

\item{padding}{padding at limits of the antigenic map, ignored if xlim or
ylim set explicitly}

\item{cex}{point size expansion factor}
}
\description{
Plot an antigenic map using the 'grid' plotting system, this is harder to
edit afterwards but allows maps with custom point shapes like "EGG" and
"UGLYEGG" to be plotted, or maps where point aspect or rotation has been
altered.
}
\seealso{
Other {functions to view maps}: 
\code{\link{RacViewer.options}()},
\code{\link{RacViewer}()},
\code{\link{export_viewer}()},
\code{\link{plot.acmap}()},
\code{\link{setLegend}()},
\code{\link{snapshotMap}()},
\code{\link{view.acmap}()},
\code{\link{view.default}()},
\code{\link{view}()}
}
\concept{{functions to view maps}}
