% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_merge_report.R
\name{htmlAdjustedTiterTable}
\alias{htmlAdjustedTiterTable}
\title{Return an html formatted titer table with antigen reactivity adjustments applied}
\usage{
htmlAdjustedTiterTable(map, optimization_number = 1)
}
\arguments{
\item{map}{An acmap object}

\item{optimization_number}{The optimization number from which to take the
antigen reactivity adjustments.}
}
\value{
A list() with a Rac_html_merge_report and shiny.tag class that can
be converted into an HTML string via as.character() and saved to a file
with save_html().
}
\description{
Prints an html formatted titer table, visualising
with colors things like which titers are the maximum for each sera.
}
