% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_merge_report.R
\name{htmlMergeReport}
\alias{htmlMergeReport}
\title{Return an html formatted merge report}
\usage{
htmlMergeReport(map)
}
\arguments{
\item{map}{An acmap object that was the result of merging several maps}
}
\value{
A list() with a Rac_html_merge_report and shiny.tag class that can
be converted into an HTML string via as.character() and saved to a file
with save_html().
}
\description{
Prints an html formatted table merge report of a set of merged maps, visualising
with colors how different titers have been merged together.
}
\seealso{
Other map merging functions: 
\code{\link{RacMerge.options}()},
\code{\link{mergeMaps}()},
\code{\link{mergeReport}()},
\code{\link{splitTiterLayers}()}
}
\concept{map merging functions}
