% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_merge_report.R
\name{htmlTiterTable}
\alias{htmlTiterTable}
\title{Return an html formatted titer table}
\usage{
htmlTiterTable(map)
}
\arguments{
\item{map}{An acmap object}
}
\value{
A list() with a Rac_html_merge_report and shiny.tag class that can
be converted into an HTML string via as.character() and saved to a file
with save_html().
}
\description{
Prints an html formatted titer table, visualising
with colors things like which titers are the maximum for each sera.
}
\seealso{
htmlAdjustedTiterTable
}
