% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{keepBestOptimization}
\alias{keepBestOptimization}
\title{Keep only the lowest stress map optimization}
\usage{
keepBestOptimization(map)
}
\arguments{
\item{map}{The acmap object}
}
\description{
Keep only the lowest stress map optimization
}
