% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{keepBestOptimization}
\alias{keepBestOptimization}
\title{Keep only the lowest stress map optimization}
\usage{
keepBestOptimization(map)
}
\arguments{
\item{map}{The acmap object}
}
\value{
An acmap object with only the lowest stress optimization kept
}
\description{
Keep only the lowest stress map optimization
}
\seealso{
Other functions for working with map data: 
\code{\link{acmap}()},
\code{\link{addOptimization}()},
\code{\link{agReactivityAdjustments}()},
\code{\link{as.json}()},
\code{\link{edit_agNames}()},
\code{\link{edit_srNames}()},
\code{\link{keepSingleOptimization}()},
\code{\link{layerNames}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{removePoints}},
\code{\link{save.acmap}()},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{subsetCommonPoints}},
\code{\link{subsetMap}()}
}
\concept{functions for working with map data}
