% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{keepSingleOptimization}
\alias{keepSingleOptimization}
\title{Keep only a single optimization run}
\usage{
keepSingleOptimization(map, optimization_number = 1)
}
\arguments{
\item{map}{The acmap object}

\item{optimization_number}{The optimization run to keep}
}
\description{
Keep only a single optimization run
}
