% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_chart.R
\name{keepSingleOptimization}
\alias{keepSingleOptimization}
\alias{keepBestOptimization}
\title{Keep a single optimization run}
\usage{
keepSingleOptimization(map, optimization_number = 1)

keepBestOptimization(map)
}
\arguments{
\item{map}{the map data object}
}
\description{
Keep only data from a single optimization run, either a specified
optimization (defaulting to the selected one), or the "best"
(lowest stress) optimization
}
\seealso{
Other {functions to work with map optimizations}: 
\code{\link{optimizationProperties}},
\code{\link{removeOptimizations}()},
\code{\link{sortOptimizations}()}
}
\concept{{functions to work with map optimizations}}
