% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{keepSingleOptimization}
\alias{keepSingleOptimization}
\title{Keep only a single optimization run}
\usage{
keepSingleOptimization(map, optimization_number = 1)
}
\arguments{
\item{map}{The acmap object}

\item{optimization_number}{The optimization run to keep}
}
\value{
An acmap object with only one optimization kept
}
\description{
Keep only a single optimization run
}
\seealso{
Other functions for working with map data: 
\code{\link{acmap}()},
\code{\link{addOptimization}()},
\code{\link{agReactivityAdjustments}()},
\code{\link{as.json}()},
\code{\link{edit_agNames}()},
\code{\link{edit_srNames}()},
\code{\link{keepBestOptimization}()},
\code{\link{layerNames}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{removePoints}},
\code{\link{save.acmap}()},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{subsetCommonPoints}},
\code{\link{subsetMap}()}
}
\concept{functions for working with map data}
