% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{layerNames}
\alias{layerNames}
\alias{layerNames<-}
\title{Get and set map layer names}
\usage{
layerNames(map)

layerNames(map) <- value
}
\arguments{
\item{map}{The acmap object}

\item{value}{A vector of new layer names to apply to the map}
}
\value{
A character vector of layer names
}
\description{
Get and set map layer names
}
\seealso{
Other functions for working with map data: 
\code{\link{acmap}()},
\code{\link{addOptimization}()},
\code{\link{agReactivityAdjustments}()},
\code{\link{as.json}()},
\code{\link{edit_agNames}()},
\code{\link{edit_srNames}()},
\code{\link{keepBestOptimization}()},
\code{\link{keepSingleOptimization}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{removePoints}},
\code{\link{save.acmap}()},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{subsetCommonPoints}},
\code{\link{subsetMap}()}
}
\concept{functions for working with map data}
