% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stress.R
\name{logtiterTable}
\alias{logtiterTable}
\title{Get the log titers from an acmap}
\usage{
logtiterTable(map)
}
\arguments{
\item{map}{The acmap object}
}
\value{
Returns a matrix of titers converted to the log scale
}
\description{
Converts titers to the log scale via via the transformation
$log2(x/10)$, lessthan values are reduced by 1 on the log scale and greater
than values are increased by 1, hence <10 => -1 and >1280 => 8
}
\seealso{
Other {map diagnostic functions}
{functions relating to map stress calculation}: 
\code{\link{mapDistances}()},
\code{\link{mapResiduals}()},
\code{\link{pointStress}},
\code{\link{recalculateStress}()},
\code{\link{stressTable}()},
\code{\link{tableDistances}()}
}
\concept{{map diagnostic functions}
{functions relating to map stress calculation}}
