% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stress.R
\name{logtiterTable}
\alias{logtiterTable}
\title{Get the log titers from an acmap}
\usage{
logtiterTable(map)
}
\arguments{
\item{map}{The acmap object}

\item{optimization_number}{The optimization data to access (defaults to the currently selected optimization)}
}
\value{
Returns a list of matrices with the logged titers and whether the titers were discrete, morethan or lessthan.
}
\description{
Get the log titers from an acmap
}
\seealso{
Other functions relating to map stress calculation: \code{\link{mapDistances}},
  \code{\link{mapResiduals}}, \code{\link{stressTable}},
  \code{\link{tableDistances}}
}
\concept{functions relating to map stress calculation}
