% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{logtiterTableLayers}
\alias{logtiterTableLayers}
\title{Return a list of logtiter table layers}
\usage{
logtiterTableLayers(map)
}
\arguments{
\item{map}{An acmap data object}
}
\description{
Return a list of logtiter table layers
}
