% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{logtiterTableLayers}
\alias{logtiterTableLayers}
\title{Return a list of logtiter table layers}
\usage{
logtiterTableLayers(map)
}
\arguments{
\item{map}{An acmap data object}
}
\value{
A list of numeric matrices with logtiter values
}
\description{
Return a list of logtiter table layers
}
\seealso{
Other map attribute functions: 
\code{\link{acmapAttributes}},
\code{\link{adjustedLogTiterTable}()},
\code{\link{adjustedTiterTable}()},
\code{\link{dilutionStepsize}()},
\code{\link{mapDescription}()},
\code{\link{mapName}()},
\code{\link{titerTableLayers}()},
\code{\link{titerTable}()}
}
\concept{map attribute functions}
