% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_bootstrap.R
\name{mapBootstrapCoords}
\alias{mapBootstrapCoords}
\alias{mapBootstrap_agCoords}
\alias{mapBootstrap_srCoords}
\title{Get bootstrap coordinates associated with a map}
\usage{
mapBootstrap_agCoords(map)

mapBootstrap_srCoords(map)
}
\arguments{
\item{map}{The map object}
}
\description{
Get bootstrap coordinates associated with a map
}
\seealso{
Other bootstrapping maps: \code{\link{bootstrapMap}}
}
\concept{bootstrapping maps}
