% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_diagnostics_bootstrap.R
\name{mapBootstrapCoords}
\alias{mapBootstrapCoords}
\alias{mapBootstrap_agCoords}
\alias{mapBootstrap_srCoords}
\title{Get bootstrap coordinates associated with a map}
\usage{
mapBootstrap_agCoords(map)

mapBootstrap_srCoords(map)
}
\arguments{
\item{map}{The map object}
}
\description{
This can be used to get information about the bootstrap run results
after \code{boostrapMap()} has been run.
}
\concept{bootstrapping maps}
