% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_diagnostics_bootstrap.R
\name{mapBootstrapCoords}
\alias{mapBootstrapCoords}
\alias{mapBootstrap_ptBaseCoords}
\alias{mapBootstrap_agCoords}
\alias{mapBootstrap_srCoords}
\title{Get bootstrap coordinates associated with a map}
\usage{
mapBootstrap_ptBaseCoords(map)

mapBootstrap_agCoords(map)

mapBootstrap_srCoords(map)
}
\arguments{
\item{map}{The map object}
}
\description{
This can be used to get information about the bootstrap run results
after \code{boostrapMap()} has been run.
}
\seealso{
Other {map diagnostic functions}: 
\code{\link{agCohesion}()},
\code{\link{bootstrapBlobs}()},
\code{\link{bootstrapMap}()},
\code{\link{checkHemisphering}()},
\code{\link{dimensionTestMap}()},
\code{\link{map-table-distances}},
\code{\link{mapRelaxed}()},
\code{\link{ptBootstrapBlob}},
\code{\link{ptBootstrapCoords}()},
\code{\link{ptLeverage}},
\code{\link{ptTriangulationBlobsize}},
\code{\link{triangulationBlobs}()},
\code{\link{unstableMaps}}
}
\concept{{map diagnostic functions}}
