% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_diagnostics_bootstrap.R
\name{mapBootstrapCoords}
\alias{mapBootstrapCoords}
\alias{mapBootstrap_ptBaseCoords}
\alias{mapBootstrap_agCoords}
\alias{mapBootstrap_srCoords}
\title{Get bootstrap coordinates associated with a map}
\usage{
mapBootstrap_ptBaseCoords(map)

mapBootstrap_agCoords(map)

mapBootstrap_srCoords(map)
}
\arguments{
\item{map}{The map object}
}
\value{
Returns a list of coordinate matrices for the points in each of
the bootstrap runs
}
\description{
This can be used to get information about the bootstrap run results
after \code{bootstrapMap()} has been run.
}
\seealso{
Other map diagnostic functions: 
\code{\link{agCohesion}()},
\code{\link{bootstrapBlobs}()},
\code{\link{bootstrapMap}()},
\code{\link{checkHemisphering}()},
\code{\link{dimensionTestMap}()},
\code{\link{logtiterTable}()},
\code{\link{map-table-distances}},
\code{\link{mapDistances}()},
\code{\link{mapRelaxed}()},
\code{\link{mapResiduals}()},
\code{\link{pointStress}},
\code{\link{ptBootstrapBlob}},
\code{\link{ptBootstrapCoords}()},
\code{\link{ptLeverage}},
\code{\link{ptTriangulationBlob}},
\code{\link{recalculateStress}()},
\code{\link{stressTable}()},
\code{\link{tableColbases}()},
\code{\link{tableDistances}()},
\code{\link{triangulationBlobs}()},
\code{\link{unstableMaps}}
}
\concept{map diagnostic functions}
