% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_optimizations.R
\name{mapComment}
\alias{mapComment}
\alias{mapComment<-}
\title{Get or set an optimization run comment}
\usage{
mapComment(map, optimization_number = 1)
mapComment(map, optimization_number = 1) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization run from which to get / set the data}

\item{value}{New value to set}
}
\value{
Gets or sets map comments for the optimization run.
}
\description{
Get or set an optimization run comment
}
\seealso{
Other map optimization attribute functions: 
\code{\link{colBases}()},
\code{\link{mapDimensions}()},
\code{\link{mapStress}()},
\code{\link{mapTransformation}()},
\code{\link{ptBaseCoords}()},
\code{\link{ptCoords}()}
}
\concept{map optimization attribute functions}
