% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{mapDescription}
\alias{mapDescription}
\alias{mapDescription<-}
\title{Getting and setting the map description}
\usage{
mapDescription(map)
mapDescription(map) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{value}{New value to set}
}
\value{
Returns either the requested attribute when using a getter function or

the updated acmap object when using the setter function.
}
\description{
Getting and setting the map description
}
\seealso{
Other map attribute functions: 
\code{\link{acmapAttributes}},
\code{\link{adjustedLogTiterTable}()},
\code{\link{adjustedTiterTable}()},
\code{\link{dilutionStepsize}()},
\code{\link{logtiterTableLayers}()},
\code{\link{mapName}()},
\code{\link{titerTableLayers}()},
\code{\link{titerTable}()}
}
\concept{map attribute functions}
