% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimizations.R
\name{mapDimensions}
\alias{mapDimensions}
\title{Get acmap dimensions}
\usage{
mapDimensions(map, optimization_number = NULL)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization data to access (defaults to the currently selected optimization)}
}
\description{
Get the number of dimensions of the selected or specified optimization.
}
\seealso{
Other functions to get and set acmap optimization attributes: \code{\link{acmapCoords}},
  \code{\link{mapComments}}, \code{\link{mapStress}},
  \code{\link{mapTransformation}},
  \code{\link{minColBasis}}
}
\concept{functions to get and set acmap optimization attributes}
