% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_optimizations.R
\name{mapDimensions}
\alias{mapDimensions}
\title{Get the current map dimensions}
\usage{
mapDimensions(map, optimization_number = 1)
}
\value{
Returns the number of map dimensions for the specified optimization run.
}
\description{
Get the current map dimensions
}
\seealso{
Other {map optimization attribute functions}: 
\code{\link{colBases}()},
\code{\link{mapComment}()},
\code{\link{mapStress}()},
\code{\link{mapTransformation}()},
\code{\link{ptBaseCoords}},
\code{\link{ptCoords}()}
}
\concept{{map optimization attribute functions}}
