% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gadget.R
\name{mapGadget}
\alias{mapGadget}
\title{Open a shiny gadget to view the map}
\usage{
mapGadget(map)
}
\arguments{
\item{map}{The acmap object to open in the GUI}
}
\description{
This function is equivalent to running \code{runGUI()} and loading a map file, but this
takes the acmap object to open as an input argument.
}
