% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gadget.R
\name{mapGadget}
\alias{mapGadget}
\title{Open a shiny gadget to view the map}
\usage{
mapGadget(map)
}
\arguments{
\item{map}{The acmap object to open in the GUI}
}
\value{
No value returned, called for the side effect of starting the gadget.
}
\description{
This function is equivalent to running \code{runGUI()} and loading a map file, but this
takes the acmap object to open as an input argument.
}
\seealso{
Other functions to view maps: 
\code{\link{RacViewer.options}()},
\code{\link{RacViewer}()},
\code{\link{export_viewer}()},
\code{\link{ggplot.acmap}()},
\code{\link{plot.acmap}()},
\code{\link{setLegend}()},
\code{\link{view.acmap}()},
\code{\link{view.default}()},
\code{\link{view}()}
}
\concept{functions to view maps}
