% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_chart.R
\name{mapInfo}
\alias{mapInfo}
\alias{sortOptimizations}
\alias{allMapStresses}
\alias{allMapDimensions}
\alias{removeOptimizations}
\alias{keepSingleOptimization}
\alias{keepBestOptimization}
\alias{set_selectedOptimization}
\title{Return information about a map}
\usage{
sortOptimizations(map)

allMapStresses(map)

allMapDimensions(map)

removeOptimizations(map)

keepSingleOptimization(map, optimization_number = NULL)

keepBestOptimization(map)

set_selectedOptimization(map, value)
}
\arguments{
\item{map}{The map object}
}
\description{
Return information about a map
}
