% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{mapName}
\alias{mapName}
\title{Getting and setting the map name}
\usage{
mapName(map)
}
\arguments{
\item{map}{The acmap data object}
}
\value{
Returns either the requested attribute when using a getter function or the updated acmap object when using the setter function.
}
\description{
You can use the standard \code{mapName()} function to get and set the map name.
}
\seealso{
Other {map attribute functions}: 
\code{\link{acmapAttributes}},
\code{\link{titerTableLayers}()},
\code{\link{titerTable}()}
}
\concept{{map attribute functions}}
