% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimize.R
\name{mapRelaxed}
\alias{mapRelaxed}
\title{Check if a map has been fully relaxed}
\usage{
mapRelaxed(map, optimization_number = NULL)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The map optimization number (defaults to the currently selected optimization)}
}
\value{
Returns TRUE or FALSE
}
\description{
Check if a map has been fully relaxed
}
