% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimize.R
\name{mapRelaxed}
\alias{mapRelaxed}
\title{Check if a map has been fully relaxed}
\usage{
mapRelaxed(map, optimization_number = 1, options = list())
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The map optimization number}

\item{options}{List of named optimizer options, see \code{RacOptimizer.options()}}
}
\value{
Returns TRUE or FALSE
}
\description{
Checks if the map optimization run can be relaxed further.
}
\seealso{
Other {map diagnostic functions}: 
\code{\link{bootstrapBlobs}()},
\code{\link{bootstrapMap}()},
\code{\link{checkHemisphering}()},
\code{\link{dimensionTestMap}()},
\code{\link{map-table-distances}},
\code{\link{mapBootstrapCoords}},
\code{\link{ptBootstrapBlob}},
\code{\link{ptBootstrapCoords}()},
\code{\link{ptLeverage}},
\code{\link{ptTriangulationBlobsize}},
\code{\link{triangulationBlobs}()}
}
\concept{{map diagnostic functions}}
