% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stress.R
\name{mapResiduals}
\alias{mapResiduals}
\title{Get a table of residuals from an acmap}
\usage{
mapResiduals(map, exclude_nd = FALSE, optimization_number = NULL)
}
\arguments{
\item{map}{The acmap object}
}
\value{
Returns a matrix of stresses, showing how much each antigen and sera
  measurement contributes to stress in the selected or specified optimization.
}
\description{
This is the difference between the table distance and the map distance
}
\seealso{
Other functions relating to map stress calculation: \code{\link{logtiterTable}},
  \code{\link{mapDistances}}, \code{\link{stressTable}},
  \code{\link{tableDistances}}
}
\concept{functions relating to map stress calculation}
