% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_antigens.R, R/map_props_chart.R,
%   R/map_props_optimizations.R, R/map_props_plotspec.R, R/map_props_sera.R
\name{mapStrains}
\alias{mapStrains}
\title{Strains in an antigenic map}
\arguments{
\item{map}{The map object}

\item{value}{Value to be used (recycled as necessary)}

\item{map}{The map object}

\item{value}{Value to be used (recycled as necessary)}

\item{map}{The map object}

\item{value}{Value to be used (recycled as necessary)}

\item{map}{The map object}

\item{value}{Value to be used (recycled as necessary)}

\item{map}{The map object}

\item{value}{Value to be used (recycled as necessary)}
}
\value{
Returns an updated racmap object

Returns an updated racmap object

Returns an updated racmap object

Returns an updated racmap object

Returns an updated racmap object
}
\description{
Strains in an antigenic map

Strains in an antigenic map

Strains in an antigenic map

Strains in an antigenic map

Strains in an antigenic map
}
