% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimizations.R
\name{mapStress}
\alias{mapStress}
\title{Get the stress associated with an acmap optimization}
\usage{
mapStress(map, optimization_number = NULL)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization data to access (defaults to the currently selected optimization)}
}
\value{
Returns the map stress
}
\description{
Gets the stress associated with the currently selected or user-specifed optimization.
}
\seealso{
See \link{pointStress} for getting the stress of individual points.

Other functions to get and set acmap optimization attributes: \code{\link{acmapCoords}},
  \code{\link{mapComments}}, \code{\link{mapDimensions}},
  \code{\link{mapTransformation}},
  \code{\link{minColBasis}}
}
\concept{functions to get and set acmap optimization attributes}
