% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_optimizations.R
\name{mapStress}
\alias{mapStress}
\title{Get the current map stress}
\usage{
mapStress(map, optimization_number = NULL)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization run to get / set the data (by default the currently selected one)}
}
\value{
Returns the current map stress value for the specified optimization run.
}
\description{
Get the current map stress
}
\seealso{
Other {map optimization attribute functions}: 
\code{\link{colBases}()},
\code{\link{mapComment}()},
\code{\link{mapDimensions}()},
\code{\link{mapTransformation}()},
\code{\link{ptBaseCoords}},
\code{\link{ptCoords}()}
}
\concept{{map optimization attribute functions}}
