% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_optimizations.R
\name{mapStress}
\alias{mapStress}
\title{Calculate the current map stress}
\usage{
mapStress(map, optimization_number = 1)
}
\arguments{
\item{map}{The acmap object}

\item{optimization_number}{The optimization number for which to calculate stress}
}
\value{
A number giving the map stress
}
\description{
Calculate the current map stress
}
\seealso{
Other map optimization attribute functions: 
\code{\link{colBases}()},
\code{\link{mapComment}()},
\code{\link{mapDimensions}()},
\code{\link{mapTransformation}()},
\code{\link{ptBaseCoords}()},
\code{\link{ptCoords}()}
}
\concept{map optimization attribute functions}
