% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimizations.R, R/map_transformation.R
\name{mapTransformation}
\alias{mapTransformation}
\alias{mapTransformation<-}
\alias{mapTranslation}
\alias{mapTranslation<-}
\alias{rotateMap}
\alias{reflectMap}
\title{Get or set acmap transformation}
\usage{
mapTransformation(map, optimization_number = NULL)

mapTransformation(map, optimization_number = NULL) <- value

mapTranslation(map, optimization_number = NULL)

mapTranslation(map, optimization_number = NULL) <- value

rotateMap(map, degrees, axis = NULL, optimization_number = NULL)

reflectMap(map, axis = "x", optimization_number = NULL)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization data to access (defaults to the currently selected optimization)}

\item{value}{Value to be used (recycled as necessary)}

\item{map}{The map object}
}
\value{
Returns an updated racmap object
}
\description{
Get or set the current transformation associated with a map.
}
\seealso{
Other functions to get and set acmap optimization attributes: \code{\link{acmapCoords}},
  \code{\link{mapComments}}, \code{\link{mapDimensions}},
  \code{\link{mapStress}}, \code{\link{minColBasis}}
}
\concept{functions to get and set acmap optimization attributes}
