% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_point_matching.R
\name{matchStrains}
\alias{matchStrains}
\alias{match_mapAntigens}
\alias{match_mapSera}
\title{Find matching antigens or sera between 2 maps}
\usage{
match_mapAntigens(map1, map2)

match_mapSera(map1, map2)
}
\arguments{
\item{map1}{The map to match names from.}

\item{map2}{The map to match names to.}
}
\value{
Returns the indices of matching strains in map 2, or NA in the
position of strains not found.
}
\description{
Find matching antigens or sera between 2 maps
}
\seealso{
Other functions to compare maps: 
\code{\link{procrustesData}()},
\code{\link{procrustesMap}()},
\code{\link{realignMap}()},
\code{\link{realignOptimizations}()}
}
\concept{functions to compare maps}
