% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_merge.R
\name{mergeReport}
\alias{mergeReport}
\title{Return a merge report}
\usage{
mergeReport(map1, map2, passage_matching = "auto")
}
\description{
Prints a raw text merge report from merging two map tables.
}
\seealso{
Other {map merging functions}: 
\code{\link{mergeMaps}()},
\code{\link{mergeTiterTables}()}
}
\concept{{map merging functions}}
