% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_merge_report.R
\name{mergeReport}
\alias{mergeReport}
\title{Return a merge report}
\usage{
mergeReport(map)
}
\arguments{
\item{map}{An acmap object that was the result of merging several maps}
}
\value{
Returns a character matrix of information on merged titers.
}
\description{
Prints a raw text merge report from merging two map tables.
}
\seealso{
Other map merging functions: 
\code{\link{RacMerge.options}()},
\code{\link{htmlMergeReport}()},
\code{\link{mergeMaps}()},
\code{\link{splitTiterLayers}()}
}
\concept{map merging functions}
