% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimize.R
\name{moveTrappedPoints}
\alias{moveTrappedPoints}
\title{Move trapped points}
\usage{
moveTrappedPoints(
  map,
  optimization_number = 1,
  stress_lim = 1,
  grid_spacing = 0.25,
  max_iterations = 10,
  options = list()
)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The map optimization number (defaults to the currently selected optimization)}
}
\value{
Returns the acmap object with updated coordinates
}
\description{
Iteratively searches for and moves points that are found to be trapped in a
optimization, stopping when no more trapped points are found.
}
\seealso{
Other {map optimization functions}: 
\code{\link{make.acmap}()},
\code{\link{optimizeMap}()},
\code{\link{randomizeCoords}()},
\code{\link{relaxMapOneStep}()},
\code{\link{relaxMap}()}
}
\concept{{map optimization functions}}
