% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.R
\name{plot.acmap}
\alias{plot.acmap}
\title{Plot an antigenic map}
\usage{
\method{plot}{acmap}(
  x,
  optimization_number = 1,
  xlim = NULL,
  ylim = NULL,
  plot_ags = TRUE,
  plot_sr = TRUE,
  plot_labels = FALSE,
  plot_blobs = TRUE,
  show_procrustes = TRUE,
  show_error_lines = TRUE,
  plot_stress = FALSE,
  indicate_outliers = TRUE,
  grid.col = "grey90",
  grid.margin.col = grid.col,
  outlier.arrow.col = grid.col,
  fill.alpha = 0.8,
  outline.alpha = 0.8,
  label.offset = 0,
  padding = 1,
  cex = 1,
  ...
)
}
\arguments{
\item{x}{The acmap to plot}

\item{optimization_number}{The optimization number to plot}

\item{xlim}{optional x axis limits}

\item{ylim}{optional y axis limits}

\item{plot_ags}{logical, should antigens be plotted}

\item{plot_sr}{logical, should antigens be plotted}

\item{plot_labels}{should point labels be plotted, can be true, false or
"antigens" or "sera"}

\item{plot_blobs}{logical, should stress blobs be plotted if present}

\item{show_procrustes}{logical, should procrustes lines be shown, if present}

\item{show_error_lines}{logical, should error lines be drawn}

\item{plot_stress}{logical, should map stress be plotted in lower left corner}

\item{indicate_outliers}{logical, should points outside of plot limits be
indicated with an arrow}

\item{grid.col}{grid line color}

\item{grid.margin.col}{grid margin color}

\item{outlier.arrow.col}{outlier arrow color}

\item{fill.alpha}{alpha for point fill}

\item{outline.alpha}{alpha for point outline}

\item{label.offset}{amount by which any point labels should be offset from
point coordinates in fractions of a character width}

\item{padding}{padding at limits of the antigenic map, ignored if xlim or
ylim set explicitly}

\item{cex}{point size expansion factor}

\item{...}{additional arguments, not used}
}
\description{
Method for plotting an antigenic map in two dimensions
}
\seealso{
Other {functions to view maps}: 
\code{\link{RacViewer.options}()},
\code{\link{RacViewer}()},
\code{\link{export_viewer}()},
\code{\link{grid.plot.acmap}()},
\code{\link{setLegend}()},
\code{\link{snapshotMap}()},
\code{\link{view.acmap}()},
\code{\link{view.default}()},
\code{\link{view}()}
}
\concept{{functions to view maps}}
