% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimizations.R
\name{pointStress}
\alias{pointStress}
\alias{agStress}
\alias{srStress}
\alias{srStressPerTiter}
\alias{agStressPerTiter}
\alias{srLikelihood}
\alias{agLikelihood}
\title{Get individual point stress}
\usage{
agStress(map, antigens = TRUE, optimization_number = NULL,
  warnings = TRUE)

srStress(map, sera = TRUE, optimization_number = NULL,
  warnings = TRUE)

srStressPerTiter(map, sera = TRUE, optimization_number = NULL)

agStressPerTiter(map, antigens = TRUE, optimization_number = NULL)

srLikelihood(map, sera = TRUE, total_error_sd,
  optimization_number = NULL, warnings = TRUE)

agLikelihood(map, antigens = TRUE, total_error_sd,
  optimization_number = NULL, warnings = TRUE)
}
\arguments{
\item{map}{The acmap data object}

\item{antigens}{Which antigens to check stress for, specified by index or name (defaults to all antigens).}

\item{optimization_number}{The optimization data to access (defaults to the currently selected optimization)}

\item{sera}{Which sera to check stress for, specified by index or name (defaults to all sera).}
}
\description{
Functions to get stress associated with individual points in a map.
}
\seealso{
See \code{\link{mapStress}} for getting the total map stress directly.
}
