% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stress.R
\name{pointStress}
\alias{pointStress}
\alias{agStress}
\alias{srStress}
\alias{srStressPerTiter}
\alias{agStressPerTiter}
\title{Get individual point stress}
\usage{
agStress(map, antigens = TRUE, optimization_number = 1, warnings = TRUE)

srStress(map, sera = TRUE, optimization_number = 1, warnings = TRUE)

srStressPerTiter(map, sera = TRUE, optimization_number = 1, exclude_nd = TRUE)

agStressPerTiter(
  map,
  antigens = TRUE,
  optimization_number = 1,
  exclude_nd = TRUE
)
}
\arguments{
\item{map}{The acmap data object}

\item{antigens}{Which antigens to check stress for, specified by index or name (defaults to all antigens).}

\item{optimization_number}{The optimization data to access (defaults to the currently selected optimization)}

\item{sera}{Which sera to check stress for, specified by index or name (defaults to all sera).}
}
\description{
Functions to get stress associated with individual points in a map.
}
\seealso{
See \code{\link{mapStress}} for getting the total map stress directly.

Other {map diagnostic functions}{functions relating to map stress calculation}: 
\code{\link{logtiterTable}()},
\code{\link{mapDistances}()},
\code{\link{mapResiduals}()},
\code{\link{recalculateStress}()},
\code{\link{stressTable}()},
\code{\link{tableDistances}()}
}
\concept{{map diagnostic functions}{functions relating to map stress calculation}}
