% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_procrustes.R
\name{procrustesMap}
\alias{procrustesMap}
\title{Return procrustes information}
\usage{
procrustesMap(
  map,
  comparison_map,
  optimization_number = 1,
  comparison_optimization_number = 1,
  translation = TRUE,
  scaling = FALSE
)

procrustesMap(
  map,
  comparison_map,
  optimization_number = 1,
  comparison_optimization_number = 1,
  translation = TRUE,
  scaling = FALSE
)
}
\arguments{
\item{map}{The acmap data object}

\item{comparison_map}{The acmap data object to procrustes against}

\item{optimization_number}{The map optimization to use in the procrustes calculation (defaults to the currently selected optimization)}

\item{translation}{Should translation be performed?}

\item{scaling}{Should scaling be performed?}

\item{antigens}{Antigens to include in the procrustes, specified either by name or index or FALSE for excluding all.}

\item{sera}{Sera to include in the procrustes, specified either by name or index or FALSE for excluding all.}

\item{target_optimization_number}{The target map optimization to use in the procrustes calculation (defaults to the currently selected optimization)}

\item{passage_matching}{Should passage matching be performed (not currently implemented)}
}
\value{
Returns procrustes information from the map to the target.

Returns procrustes information from the map to the target.
}
\description{
Returns information from one map procrusted to another.

Returns information from one map procrusted to another.
}
\seealso{
Other {functions to compare maps}: 
\code{\link{realignMap}()},
\code{\link{realignOptimizations}()}

Other {functions to compare maps}: 
\code{\link{realignMap}()},
\code{\link{realignOptimizations}()}
}
\concept{{functions to compare maps}}
