% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_procrustes.R
\name{procrustesMap}
\alias{procrustesMap}
\title{Return procrustes information}
\usage{
procrustesMap(map, target_map, antigens = TRUE, sera = TRUE,
  translation = TRUE, scaling = FALSE, optimization_number = NULL,
  target_optimization_number = NULL, passage_matching = "ignore")
}
\arguments{
\item{map}{The acmap data object}

\item{target_map}{The acmap data object to procrustes against}

\item{antigens}{Antigens to include in the procrustes, specified either by name or index or FALSE for excluding all.}

\item{sera}{Sera to include in the procrustes, specified either by name or index or FALSE for excluding all.}

\item{translation}{Should translation be performed?}

\item{scaling}{Should scaling be performed?}

\item{optimization_number}{The map optimization to use in the procrustes calculation (defaults to the currently selected optimization)}

\item{target_optimization_number}{The target map optimization to use in the procrustes calculation (defaults to the currently selected optimization)}

\item{passage_matching}{Should passage matching be performed}
}
\value{
Returns procrustes information from the map to the target.
}
\description{
Returns information from one map procrusted to another.
}
