% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_antigens.R
\name{ptAnnotations}
\alias{ptAnnotations}
\alias{agAnnotations}
\alias{srAnnotations}
\alias{agAnnotations<-}
\alias{srAnnotations<-}
\title{Getting and setting point annotation information}
\usage{
agAnnotations(map)

srAnnotations(map)

agAnnotations(map) <- value

srAnnotations(map) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{value}{A list of character vectors with annotations information for each
point}
}
\value{
A character vector of point annotations.
}
\description{
Getting and setting point annotation information
}
\seealso{
Other antigen and sera attribute functions: 
\code{\link{agAttributes}},
\code{\link{agGroups}()},
\code{\link{agHomologousSr}()},
\code{\link{agLabIDs}()},
\code{\link{agSequences}()},
\code{\link{ptClades}},
\code{\link{srAttributes}},
\code{\link{srGroups}()},
\code{\link{srHomologousAgs}()},
\code{\link{srSequences}()}
}
\concept{antigen and sera attribute functions}
