% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_optimizations.R
\name{ptBaseCoords}
\alias{ptBaseCoords}
\alias{agBaseCoords}
\alias{srBaseCoords}
\title{Getting and setting base coordinates}
\usage{
agBaseCoords(map, optimization_number = NULL)

srBaseCoords(map, optimization_number = NULL)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization run to get / set the data (by default the currently selected one)}
}
\value{
Returns either the requested attribute when using a getter function or the updated acmap object when using the setter function.
}
\description{
These functions get and set the base coordinates for a given optimization run.
}
\seealso{
\code{\link{agCoords}}
\code{\link{srCoords}}

Other {map optimization attribute functions}: 
\code{\link{colBases}()},
\code{\link{mapComment}()},
\code{\link{mapDimensions}()},
\code{\link{mapStress}()},
\code{\link{mapTransformation}()},
\code{\link{ptCoords}()}
}
\concept{{map optimization attribute functions}}
