% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_optimizations.R
\name{ptBaseCoords}
\alias{ptBaseCoords}
\alias{agBaseCoords}
\alias{agBaseCoords<-}
\alias{srBaseCoords}
\alias{srBaseCoords<-}
\title{Getting and setting base coordinates}
\usage{
ptBaseCoords(map, optimization_number = 1)
agBaseCoords(map, optimization_number = 1)
agBaseCoords(map, optimization_number = 1) <- value
srBaseCoords(map, optimization_number = 1)
srBaseCoords(map, optimization_number = 1) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization run from which to get / set the data}

\item{value}{New value to set}
}
\value{
Returns either the requested attribute when using a getter function or

the updated acmap object when using the setter function.
}
\description{
These functions get and set the base coordinates for a given optimization
run.
}
\seealso{
\code{agCoords()}
\code{srCoords()}

Other map optimization attribute functions: 
\code{\link{colBases}()},
\code{\link{mapComment}()},
\code{\link{mapDimensions}()},
\code{\link{mapStress}()},
\code{\link{mapTransformation}()},
\code{\link{ptCoords}()}
}
\concept{map optimization attribute functions}
