% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_diagnostics_bootstrap.R
\name{ptBootstrapBlob}
\alias{ptBootstrapBlob}
\alias{agBootstrapBlob}
\alias{srBootstrapBlob}
\title{Get antigen or serum blob information}
\usage{
agBootstrapBlob(map, antigen)

srBootstrapBlob(map, serum)
}
\arguments{
\item{map}{An acmap object}

\item{antigen}{The antigen to get the blob for}

\item{serum}{The serum to get the blob for}
}
\value{
Returns an object of class "blob" that can be plotted using the \code{blob()} funciton.
}
\description{
Get antigen or serum blob information for plotting with the \code{blob()} function.
}
\seealso{
Other {map diagnostic functions}: 
\code{\link{bootstrapBlobs}()},
\code{\link{bootstrapMap}()},
\code{\link{checkHemisphering}()},
\code{\link{dimensionTestMap}()},
\code{\link{map-table-distances}},
\code{\link{mapBootstrapCoords}},
\code{\link{mapRelaxed}()},
\code{\link{ptBootstrapCoords}()},
\code{\link{ptLeverage}},
\code{\link{ptTriangulationBlobsize}},
\code{\link{triangulationBlobs}()}
}
\concept{{map diagnostic functions}}
