% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_diagnostics_bootstrap.R
\name{ptBootstrapCoords}
\alias{ptBootstrapCoords}
\alias{agBootstrapCoords}
\alias{srBootstrapCoords}
\title{Get antigen or serum bootstrap coordinates information}
\usage{
ptBootstrapCoords(map, point)

agBootstrapCoords(map, antigen)

srBootstrapCoords(map, serum)
}
\arguments{
\item{map}{An acmap object}

\item{point}{The point from which to get the bootstrap coords (numbered
antigens then sera)}

\item{antigen}{The antigen to get the bootstrap coords}

\item{serum}{The serum from which to get the bootstrap coords}
}
\value{
Returns a matrix of coordinates for the point in each of the
bootstrap runs
}
\description{
Get antigen or serum bootstrap coordinates information
}
\seealso{
Other map diagnostic functions: 
\code{\link{agCohesion}()},
\code{\link{bootstrapBlobs}()},
\code{\link{bootstrapMap}()},
\code{\link{checkHemisphering}()},
\code{\link{dimensionTestMap}()},
\code{\link{logtiterTable}()},
\code{\link{map-table-distances}},
\code{\link{mapBootstrapCoords}},
\code{\link{mapDistances}()},
\code{\link{mapRelaxed}()},
\code{\link{mapResiduals}()},
\code{\link{pointStress}},
\code{\link{ptBootstrapBlob}},
\code{\link{ptLeverage}},
\code{\link{ptTriangulationBlob}},
\code{\link{recalculateStress}()},
\code{\link{stressTable}()},
\code{\link{tableColbases}()},
\code{\link{tableDistances}()},
\code{\link{triangulationBlobs}()},
\code{\link{unstableMaps}}
}
\concept{map diagnostic functions}
