% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_blobs.R
\name{ptStressBlobSize}
\alias{ptStressBlobSize}
\alias{agStressBlobSize}
\alias{srStressBlobSize}
\title{Fetch information on stress blob size}
\usage{
agStressBlobSize(map)

srStressBlobSize(map)
}
\arguments{
\item{map}{acmap with stress blob information added}
}
\description{
Returns a vector of stress blob sizes for each point, helpful for
programatically finding the points with the most uncertainty.
}
