% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_plotspec.R
\name{ptStyles}
\alias{ptStyles}
\alias{agShown}
\title{Getting and setting point plotting styles}
\usage{
agShown(map)
}
\description{
These functions get and set the styles to use for each point when plotting.
}
\seealso{
Other {map point style functions}: 
\code{\link{applyPlotspec}()}
}
\concept{{map point style functions}}
