% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_plotspec.R
\name{ptStyles}
\alias{ptStyles}
\alias{agShown}
\alias{srShown}
\alias{agShown<-}
\alias{srShown<-}
\alias{agSize}
\alias{srSize}
\alias{agSize<-}
\alias{srSize<-}
\alias{agFill}
\alias{srFill}
\alias{agFill<-}
\alias{srFill<-}
\alias{agOutline}
\alias{srOutline}
\alias{agOutline<-}
\alias{srOutline<-}
\alias{agOutlineWidth}
\alias{srOutlineWidth}
\alias{agOutlineWidth<-}
\alias{srOutlineWidth<-}
\alias{agRotation}
\alias{srRotation}
\alias{agRotation<-}
\alias{srRotation<-}
\alias{agAspect}
\alias{srAspect}
\alias{agAspect<-}
\alias{srAspect<-}
\alias{agShape}
\alias{srShape}
\alias{agShape<-}
\alias{srShape<-}
\title{Getting and setting point plotting styles}
\usage{
agShown(map)

srShown(map)

agShown<-(map)

srShown<-(map)

agSize(map)

srSize(map)

agSize<-(map)

srSize<-(map)

agFill(map)

srFill(map)

agFill<-(map)

srFill<-(map)

agOutline(map)

srOutline(map)

agOutline<-(map)

srOutline<-(map)

agOutlineWidth(map)

srOutlineWidth(map)

agOutlineWidth<-(map)

srOutlineWidth<-(map)

agRotation(map)

srRotation(map)

agRotation<-(map)

srRotation<-(map)

agAspect(map)

srAspect(map)

agAspect<-(map)

srAspect<-(map)

agShape(map)

srShape(map)

agShape<-(map)

srShape<-(map)
}
\arguments{
\item{map}{The acmap data object}
}
\value{
Returns either the requested attribute when using a getter function or the updated acmap object when using the setter function.
}
\description{
These functions get and set the styles to use for each point when plotting.
}
\seealso{
Other {map point style functions}: 
\code{\link{applyPlotspec}()}
}
\concept{{map point style functions}}
