% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_plotspec.R
\name{ptStyles}
\alias{ptStyles}
\alias{agShown}
\alias{srShown}
\alias{agShown<-}
\alias{srShown<-}
\alias{agSize}
\alias{srSize}
\alias{agSize<-}
\alias{srSize<-}
\alias{agFill}
\alias{srFill}
\alias{agFill<-}
\alias{srFill<-}
\alias{agOutline}
\alias{srOutline}
\alias{agOutline<-}
\alias{srOutline<-}
\alias{agOutlineWidth}
\alias{srOutlineWidth}
\alias{agOutlineWidth<-}
\alias{srOutlineWidth<-}
\alias{agRotation}
\alias{srRotation}
\alias{agRotation<-}
\alias{srRotation<-}
\alias{agAspect}
\alias{srAspect}
\alias{agAspect<-}
\alias{srAspect<-}
\alias{agShape}
\alias{srShape}
\alias{agShape<-}
\alias{srShape<-}
\title{Getting and setting point plotting styles}
\usage{
agShown(map)
srShown(map)
agShown(map) <- value
srShown(map) <- value
agSize(map)
srSize(map)
agSize(map) <- value
srSize(map) <- value
agFill(map)
srFill(map)
agFill(map) <- value
srFill(map) <- value
agOutline(map)
srOutline(map)
agOutline(map) <- value
srOutline(map) <- value
agOutlineWidth(map)
srOutlineWidth(map)
agOutlineWidth(map) <- value
srOutlineWidth(map) <- value
agRotation(map)
srRotation(map)
agRotation(map) <- value
srRotation(map) <- value
agAspect(map)
srAspect(map)
agAspect(map) <- value
srAspect(map) <- value
agShape(map)
srShape(map)
agShape(map) <- value
srShape(map) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{value}{New value to set}
}
\value{
Returns either the requested attribute when using a getter function or

the updated acmap object when using the setter function.
}
\description{
These functions get and set the styles to use for each point when plotting.
}
\seealso{
Other map point style functions: 
\code{\link{applyPlotspec}()},
\code{\link{ptDrawingOrder}()},
\code{\link{ptOpacity}}
}
\concept{map point style functions}
