% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimize.R
\name{RacOptimizer.options}
\alias{RacOptimizer.options}
\title{Set acmap optimization options}
\usage{
RacOptimizer.options(
  dim_annealing = FALSE,
  method = "L-BFGS",
  maxit = 1000,
  num_cores = parallel::detectCores(),
  report_progress = TRUE,
  progress_bar_length = options()$width
)
}
\arguments{
\item{point.opacity}{Default opacity for unselected points}

\item{viewer.controls}{Should viewer controls be shown or hidden by default?}
}
\value{
Returns a named list of optimizer options
}
\description{
This function facilitates setting options for the acmap optimizer process by returning a list of option settings.
}
