% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimize.R
\name{RacOptimizer.options}
\alias{RacOptimizer.options}
\title{Set acmap optimization options}
\usage{
RacOptimizer.options(
  dim_annealing = FALSE,
  method = "L-BFGS",
  maxit = 1000,
  num_cores = parallel::detectCores(),
  report_progress = NULL,
  ignore_disconnected = FALSE,
  progress_bar_length = options()$width
)
}
\arguments{
\item{dim_annealing}{Should dimensional annealing be performed}

\item{method}{The optimization method to use}

\item{maxit}{The maximum number of iterations to use in the optimizer}

\item{num_cores}{The number of cores to run in parallel}

\item{report_progress}{Should progress be reported}

\item{ignore_disconnected}{Should the check for disconnected points be skipped}

\item{progress_bar_length}{Progress bar length when progress is reported}
}
\value{
Returns a named list of optimizer options
}
\description{
This function facilitates setting options for the acmap optimizer process by
returning a list of option settings.
}
\details{
For more details, for example on "dimensional annealing" see
\code{vignette("intro-to-antigenic-cartography")}. For details on optimizer
settings like \code{maxit} see the underlying optimizer documentation at
\href{http://ensmallen.org}{ensmallen.org}.
}
\seealso{
Other map optimization functions: 
\code{\link{make.acmap}()},
\code{\link{moveTrappedPoints}()},
\code{\link{optimizeMap}()},
\code{\link{randomizeCoords}()},
\code{\link{relaxMapOneStep}()},
\code{\link{relaxMap}()}
}
\concept{map optimization functions}
