% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_viewer.R
\name{RacViewer.options}
\alias{RacViewer.options}
\title{Set viewer options}
\usage{
RacViewer.options(
  point.opacity = NA,
  viewer.controls = "hidden",
  grid.display = "static"
)
}
\arguments{
\item{point.opacity}{Default opacity for unselected points}

\item{viewer.controls}{Should viewer controls be shown or hidden by default?}
}
\value{
Returns a named list of viewer options
}
\description{
This function facilitates setting racviewer options by returning a list of option settings.
}
