% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RacViewer.R
\name{RacViewer}
\alias{RacViewer}
\title{Create a RacViewer widget}
\usage{
RacViewer(map, rotation, translation, zoom, plotdata = NULL,
  show_procrustes = FALSE, hide_control_panel = FALSE, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{map}{The map data object}

\item{rotation}{Viewer rotation}

\item{translation}{Viewer translation}

\item{zoom}{Viewer zoom}

\item{width}{Width of the widget}

\item{height}{Height of the widget}

\item{elementId}{DOM element ID}

\item{controls}{Should controls be shown}

\item{stress}{Should stress be shown}

\item{show_ags}{Should antigens be shown}

\item{show_sr}{Should sera be shown}

\item{selectable}{Should points be selectable}
}
\description{
This creates an html widget for viewing antigenic maps.
}
