% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RacViewer.R
\name{RacViewer}
\alias{RacViewer}
\title{Create a RacViewer widget}
\usage{
RacViewer(
  map,
  plotdata = NULL,
  show_procrustes = FALSE,
  options = list(),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{map}{The map data object}

\item{plotdata}{r3js plot data}

\item{show_procrustes}{should procrustes lines be shown}

\item{options}{A named list of viewer options supplied to
\code{racviewer.options()}}

\item{width}{Width of the widget}

\item{height}{Height of the widget}

\item{elementId}{DOM element ID}
}
\description{
This creates an html widget for viewing antigenic maps.
}
\seealso{
Other {functions to view maps}: 
\code{\link{RacViewer.options}()},
\code{\link{export_viewer}()},
\code{\link{grid.plot.acmap}()},
\code{\link{mapGadget}()},
\code{\link{plot.acmap}()},
\code{\link{setLegend}()},
\code{\link{snapshotMap}()},
\code{\link{view.default}()},
\code{\link{view}()}
}
\concept{{functions to view maps}}
