% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_process_titers.R
\name{read.titerTable}
\alias{read.titerTable}
\title{Read in a table of titer data}
\usage{
read.titerTable(filepath)
}
\arguments{
\item{filepath}{Path to the table of titer data}
}
\value{
Returns a matrix of titers.
}
\description{
Reads in a table of titer data, converting it to a matrix of titers with
labelled column and row names. Missing titers should be represented by an
asterisk character.
}
\details{
Currently supported file formats are .csv and .xls and .txt
}
\seealso{
Other functions for working with map data: 
\code{\link{acmap}()},
\code{\link{addOptimization}()},
\code{\link{agReactivityAdjustments}()},
\code{\link{as.json}()},
\code{\link{edit_agNames}()},
\code{\link{edit_srNames}()},
\code{\link{keepBestOptimization}()},
\code{\link{keepSingleOptimization}()},
\code{\link{layerNames}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{removePoints}},
\code{\link{save.acmap}()},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{subsetCommonPoints}},
\code{\link{subsetMap}()}
}
\concept{functions for working with map data}
