% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_procrustes.R
\name{realignOptimizations}
\alias{realignOptimizations}
\title{Realigns optimizations in the map}
\usage{
realignOptimizations(map)
}
\arguments{
\item{map}{The acmap data object}
}
\value{
Returns the map with realigned optimizations
}
\description{
Realigns all map optimizations through rotatation and translation to match
point positions as closely as possible to the first optimization run. This
is done by default when optimizing a map and makes comparing point positions
in each optimization run much easier to do by eye.
}
\seealso{
Other functions to compare maps: 
\code{\link{matchStrains}},
\code{\link{procrustesData}()},
\code{\link{procrustesMap}()},
\code{\link{realignMap}()}
}
\concept{functions to compare maps}
