% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stress.R
\name{recalculateStress}
\alias{recalculateStress}
\title{Recalculate the stress associated with an acmap optimization}
\usage{
recalculateStress(map, optimization_number = NULL)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization data to access (defaults to the currently selected optimization)}
}
\value{
Returns the map stress
}
\description{
Recalculates the stress associated with the currently selected or user-specifed optimization.
}
\seealso{
See \link{pointStress} for getting the stress of individual points.

Other {map diagnostic functions}{functions relating to map stress calculation}: 
\code{\link{logtiterTable}()},
\code{\link{mapDistances}()},
\code{\link{mapResiduals}()},
\code{\link{pointStress}},
\code{\link{stressTable}()},
\code{\link{tableDistances}()}
}
\concept{{map diagnostic functions}{functions relating to map stress calculation}}
