% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_optimize.R
\name{relaxMapOneStep}
\alias{relaxMapOneStep}
\title{Relax a map one step in the optimiser}
\usage{
relaxMapOneStep(
  map,
  optimization_number = 1,
  fixed_antigens = FALSE,
  fixed_sera = FALSE,
  options = list()
)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The map optimization number}

\item{fixed_antigens}{Antigens to set fixed positions for when relaxing}

\item{fixed_sera}{Sera to set fixed positions for when relaxing}

\item{options}{List of named optimizer options, see \code{RacOptimizer.options()}}
}
\value{
Returns an updated map object
}
\description{
Relax a map one step in the optimiser
}
\seealso{
Other map optimization functions: 
\code{\link{RacOptimizer.options}()},
\code{\link{make.acmap}()},
\code{\link{moveTrappedPoints}()},
\code{\link{optimizeMap}()},
\code{\link{randomizeCoords}()},
\code{\link{relaxMap}()}
}
\concept{map optimization functions}
