% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{removeOptimizations}
\alias{removeOptimizations}
\title{Remove map optimizations}
\usage{
removeOptimizations(map)
}
\arguments{
\item{map}{The acmap object}
}
\value{
An acmap object with all optimizations removed
}
\description{
Remove all optimization run data from a map object
}
\seealso{
Other functions to work with map optimizations: 
\code{\link{keepOptimizations}()},
\code{\link{optimizationProperties}},
\code{\link{sortOptimizations}()}
}
\concept{functions to work with map optimizations}
