% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_subset.R
\name{removePoints}
\alias{removePoints}
\alias{removeAntigens}
\alias{removeSera}
\title{Remove antigens and sera}
\usage{
removeAntigens(map, antigens)

removeSera(map, sera)
}
\arguments{
\item{map}{The map data object}

\item{antigens}{Antigens to remove (specified by name or index)}

\item{sera}{Sera to remove (specified by name or index)}
}
\value{
An acmap object with points removed
}
\description{
Functions to remove antigens and sera from a map
}
\seealso{
Other functions for working with map data: 
\code{\link{acmap}()},
\code{\link{addOptimization}()},
\code{\link{agReactivityAdjustments}()},
\code{\link{as.json}()},
\code{\link{edit_agNames}()},
\code{\link{edit_srNames}()},
\code{\link{keepBestOptimization}()},
\code{\link{keepSingleOptimization}()},
\code{\link{layerNames}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{save.acmap}()},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{subsetCommonPoints}},
\code{\link{subsetMap}()}
}
\concept{functions for working with map data}
