% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_transformation.R
\name{rotateMap}
\alias{rotateMap}
\title{Rotate a map}
\usage{
rotateMap(map, degrees, axis = NULL, optimization_number = NULL)
}
\arguments{
\item{map}{The acmap object}

\item{degrees}{Degrees of rotation}

\item{axis}{Axis of rotation (if 3D), specified as "x", "y", or "z"}

\item{optimization_number}{The optimization number (or NULL to apply to all
optimizations)}
}
\value{
An acmap object with rotation applied
}
\description{
Apply a rotation to an antigenic map
}
\seealso{
Other functions relating to map transformation: 
\code{\link{applyMapTransform}()},
\code{\link{reflectMap}()},
\code{\link{translateMap}()}
}
\concept{functions relating to map transformation}
