% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_loadsave.R
\name{save.acmap}
\alias{save.acmap}
\title{Save acmap data to a file}
\usage{
save.acmap(map, filename)
}
\arguments{
\item{map}{The acmap data object.}

\item{filename}{Path to the file.}
}
\description{
Save acmap data to a file in a given format, must be one of '.ace', '.save' or '.save.xs' (see details).
}
\details{
Although several file formats are supported, you should save files
  with the '.ace' extension. This is the most recent, while the others are
  mostly there for backwards compatibility to other software.
}
\seealso{
Other functions to create and save acmap objects: \code{\link{acmap}},
  \code{\link{read.acmap.cpp}}, \code{\link{read.acmap}},
  \code{\link{save.coords}}, \code{\link{save.titerTable}}
}
\concept{functions to create and save acmap objects}
