% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_loadsave.R
\name{save.acmap}
\alias{save.acmap}
\title{Save acmap data to a file}
\usage{
save.acmap(map, filename)
}
\arguments{
\item{map}{The acmap data object.}

\item{filename}{Path to the file.}
}
\description{
Save acmap data to a file. The preferred extension is ".ace", although
the format of the file will be a json file of map data compressed using
'xz' compression.
}
\seealso{
Other {functions for working with map data}: 
\code{\link{acmap}()},
\code{\link{as.json}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{removePoints}},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{subsetMap}()}
}
\concept{{functions for working with map data}}
