% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_loadsave.R
\name{save.coords}
\alias{save.coords}
\title{Save acmap coordinate data to a file}
\usage{
save.coords(map, filename, antigens = TRUE, sera = TRUE)
}
\arguments{
\item{map}{The acmap data object.}

\item{filename}{Path to the file.}
}
\description{
Saves acmap coordinate data of all or specified antigens and sera to a .csv file.
}
\seealso{
Other {functions for working with map data}: 
\code{\link{acmap}()},
\code{\link{as.cpp}()},
\code{\link{as.json}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{removePoints}},
\code{\link{save.acmap}()},
\code{\link{save.titerTable}()},
\code{\link{selectedOptimization}()},
\code{\link{subsetMap}()}
}
\concept{{functions for working with map data}}
