% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{sortOptimizations}
\alias{sortOptimizations}
\title{Sort optimizations by stress}
\usage{
sortOptimizations(map)
}
\arguments{
\item{map}{The acmap object}
}
\value{
An acmap object with optimizations sorted by stress.
}
\description{
Sorts all the optimization runs for a given map object by stress
(lowest to highest). Note that this is done by default when running
\code{optimizeMap()}.
}
\seealso{
Other functions to work with map optimizations: 
\code{\link{keepOptimizations}()},
\code{\link{optimizationProperties}},
\code{\link{removeOptimizations}()}
}
\concept{functions to work with map optimizations}
