% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_sera.R
\name{srAttributes}
\alias{srAttributes}
\alias{srIDs}
\alias{srNames}
\alias{srNames<-}
\alias{srExtra}
\alias{srExtra<-}
\alias{srIDs<-}
\alias{srDates}
\alias{srDates<-}
\alias{srNamesFull}
\alias{srNamesAbbreviated}
\alias{srReference}
\alias{srPassage}
\title{Getting and setting sera attributes}
\usage{
srNames(map)
srNames(map) <- value
srExtra(map)
srExtra(map) <- value
srIDs(map)
srIDs(map) <- value
srDates(map)
srDates(map) <- value
srNamesFull(map)
srNamesAbbreviated(map)
srReference(map)
srPassage(map)
}
\arguments{
\item{map}{The acmap data object}

\item{value}{New value to set}
}
\value{
Returns either the requested attribute when using a getter function or

the updated acmap object when using the setter function.
}
\description{
These functions get and set the sera attributes for a map.
}
\seealso{
\code{agAttributes()}

Other {antigen and sera attribute functions}: 
\code{\link{agAttributes}},
\code{\link{agGroups}()},
\code{\link{agLabIDs}()},
\code{\link{agSequences}()},
\code{\link{ptAnnotations}},
\code{\link{ptClades}},
\code{\link{srGroups}()},
\code{\link{srSequences}()}
}
\concept{{antigen and sera attribute functions}}
