% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_sera.R
\name{srAttributes}
\alias{srAttributes}
\alias{srIDs}
\alias{srNames}
\alias{srNames<-}
\alias{srIDs<-}
\alias{srNamesFull}
\alias{srNamesAbbreviated}
\alias{srDates}
\alias{srDates<-}
\alias{srReference}
\title{Getting and setting sera attributes}
\usage{
srNames(map)

srNames<-(map)

srIDs(map)

srIDs<-(map)

srNamesFull(map)

srNamesAbbreviated(map)

srDates(map)

srDates<-(map)

srReference(map)
}
\arguments{
\item{map}{The acmap data object}
}
\value{
Returns either the requested attribute when using a getter function or the updated acmap object when using the setter function.
}
\description{
These functions get and set the sera attributes for a map.
}
\seealso{
\code{\link{agAttributes}}

Other {antigen and sera attribute functions}: 
\code{\link{agAttributes}},
\code{\link{agGroups}()},
\code{\link{srGroups}()}
}
\concept{{antigen and sera attribute functions}}
