% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_sera.R
\name{srAttributes}
\alias{srAttributes}
\alias{srIDs}
\alias{srNames}
\alias{srGroups}
\alias{srNamesFull}
\alias{srNamesAbbreviated}
\title{Getting and setting sera attributes}
\usage{
srNames(map, .name = TRUE)
srNames(map, .check = TRUE) <- value

srIDs(map, .name = TRUE)
srIDs(map, .check = TRUE) <- value

srGroups(map, .name = TRUE)
srGroups(map, .check = TRUE) <- value

srNamesFull(map, .name = TRUE)

srNamesAbbreviated(map, .name = TRUE)
}
\arguments{
\item{map}{The acmap data object}

\item{.name}{Should the output be labelled with corresponding antigen / sera names}

\item{.check}{Should normal checks be applied for validity be applied when setting this value (used internally)}
}
\value{
Returns either the requested attribute when using a getter function or the updated acmap object when using the setter function.
}
\description{
These functions get and set the sera attributes for a given optimization run.
}
\seealso{
\code{\link{agAttributes}}

Other {antigen and sera attribute functions}: 
\code{\link{agAttributes}}
}
\concept{{antigen and sera attribute functions}}
