% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_sera.R
\name{srAttributes}
\alias{srAttributes}
\alias{srIDs}
\alias{srIDs<-}
\alias{srDates}
\alias{srDates<-}
\alias{srReference}
\alias{srReference<-}
\alias{srNames}
\alias{srNames<-}
\alias{srExtra}
\alias{srExtra<-}
\alias{srPassage}
\alias{srPassage<-}
\alias{srLineage}
\alias{srLineage<-}
\alias{srReassortant}
\alias{srReassortant<-}
\alias{srStrings}
\alias{srStrings<-}
\alias{srSpecies}
\alias{srSpecies<-}
\title{Getting and setting sera attributes}
\usage{
srIDs(map)
srIDs(map) <- value
srDates(map)
srDates(map) <- value
srReference(map)
srReference(map) <- value
srNames(map)
srNames(map) <- value
srExtra(map)
srExtra(map) <- value
srPassage(map)
srPassage(map) <- value
srLineage(map)
srLineage(map) <- value
srReassortant(map)
srReassortant(map) <- value
srStrings(map)
srStrings(map) <- value
srSpecies(map)
srSpecies(map) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{value}{New value to set}
}
\value{
Returns either the requested attribute when using a getter function or

the updated acmap object when using the setter function.
}
\description{
These functions get and set the sera attributes for a map.
}
\seealso{
\code{agAttributes()}

Other antigen and sera attribute functions: 
\code{\link{agAttributes}},
\code{\link{agGroups}()},
\code{\link{agHomologousSr}()},
\code{\link{agLabIDs}()},
\code{\link{agSequences}()},
\code{\link{ptAnnotations}},
\code{\link{ptClades}},
\code{\link{srGroups}()},
\code{\link{srHomologousAgs}()},
\code{\link{srSequences}()}
}
\concept{antigen and sera attribute functions}
