% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_sera.R
\name{srGroups}
\alias{srGroups}
\alias{srGroups<-}
\title{Getting and setting sera groups}
\usage{
srGroups(map)

srGroups(map) <- value
}
\arguments{
\item{map}{The acmap object}

\item{value}{A character or factor vector of groupings to apply to the sera}
}
\value{
A factor vector of serum groups
}
\description{
These functions get and set the sera groupings for a map.
}
\seealso{
Other antigen and sera attribute functions: 
\code{\link{agAttributes}},
\code{\link{agGroups}()},
\code{\link{agHomologousSr}()},
\code{\link{agLabIDs}()},
\code{\link{agSequences}()},
\code{\link{ptAnnotations}},
\code{\link{ptClades}},
\code{\link{srAttributes}},
\code{\link{srHomologousAgs}()},
\code{\link{srSequences}()}
}
\concept{antigen and sera attribute functions}
