% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_sera.R
\name{srSequences}
\alias{srSequences}
\alias{srSequences<-}
\title{Getting and setting sera sequence information}
\usage{
srSequences(map)

srSequences(map) <- value
}
\arguments{
\item{map}{The acmap data object}
}
\description{
Getting and setting sera sequence information
}
