% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props_points_sera.R
\name{srSequences}
\alias{srSequences}
\alias{srSequences<-}
\title{Getting and setting sera sequence information}
\usage{
srSequences(map)

srSequences(map) <- value
}
\arguments{
\item{map}{The acmap data object}

\item{value}{A character matrix of sequences with rows equal to the number of
sera}
}
\description{
Getting and setting sera sequence information
}
\seealso{
Other {antigen and sera attribute functions}: 
\code{\link{agAttributes}},
\code{\link{agGroups}()},
\code{\link{agSequences}()},
\code{\link{srAttributes}},
\code{\link{srGroups}()}
}
\concept{{antigen and sera attribute functions}}
