% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stress.R
\name{stressTable}
\alias{stressTable}
\title{Get a stress table from an acmap}
\usage{
stressTable(map, optimization_number = NULL)
}
\arguments{
\item{map}{The acmap object}
}
\value{
Returns a matrix of stresses, showing how much each antigen and sera
  measurement contributes to stress in the selected or specified optimization.
}
\description{
Get a stress table from an acmap
}
\seealso{
Other {map diagnostic functions}{functions relating to map stress calculation}: 
\code{\link{logtiterTable}()},
\code{\link{mapDistances}()},
\code{\link{mapResiduals}()},
\code{\link{pointStress}},
\code{\link{recalculateStress}()},
\code{\link{tableDistances}()}
}
\concept{{map diagnostic functions}{functions relating to map stress calculation}}
