% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_subset.R
\name{subsetCommonPoints}
\alias{subsetCommonPoints}
\alias{subsetCommonAgs}
\alias{subsetCommonSrGroups}
\title{Remove antigens and sera}
\usage{
subsetCommonAgs(maps)

subsetCommonSrGroups(maps)
}
\arguments{
\item{maps}{A list of map data objects}
}
\description{
Functions to subset a list of maps to include only antigens, antigen groups, sera
or serum groups that are in common between them.
}
