% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_subset.R
\name{subsetCommonPoints}
\alias{subsetCommonPoints}
\alias{subsetCommonAgs}
\alias{subsetCommonSrGroups}
\title{Remove antigens and sera}
\usage{
subsetCommonAgs(maps)

subsetCommonSrGroups(maps)
}
\arguments{
\item{maps}{A list of map data objects}
}
\description{
Functions to subset a list of maps to include only antigens, antigen groups, sera
or serum groups that are in common between them.
}
\seealso{
Other functions for working with map data: 
\code{\link{acmap}()},
\code{\link{addOptimization}()},
\code{\link{agReactivityAdjustments}()},
\code{\link{as.json}()},
\code{\link{edit_agNames}()},
\code{\link{edit_srNames}()},
\code{\link{keepBestOptimization}()},
\code{\link{keepSingleOptimization}()},
\code{\link{layerNames}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{removePoints}},
\code{\link{save.acmap}()},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{subsetMap}()}
}
\concept{functions for working with map data}
