% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_subset.R
\name{subsetMap}
\alias{subsetMap}
\title{Subset an antigenic map}
\usage{
subsetMap(map, antigens = TRUE, sera = TRUE)
}
\arguments{
\item{map}{The antigenic map object}

\item{antigens}{Antigens to keep}

\item{sera}{Sera to keep}
}
\value{
Returns a new antigenic map containing only match antigens and sera
}
\description{
Subset an antigenic map to contain only specified antigens and sera
}
\examples{

# Subsetting an antigenic map -------

## Subset the map by antigen and sera indices
map_subset1 <- subsetMap(h3map2004,
                         antigens = 1:4,
                         sera     = 1:4)

## Subset the map by antigen and sera names
map_subset2 <- subsetMap(h3map2004,
                         antigens = c("BI/2600/75", "HK/3/95"),
                         sera     = c("HK/34/90", "NL/172/96"))

## If nothing is given for the antigen or sera argument then the default is to select all
map_subset3 <- subsetMap(h3map2004,
                         antigens = c("BI/2600/75", "HK/3/95"))

## A warning is returned if an antigen or sera are not found
map_subset4 <- subsetMap(h3map2004,
                         antigens = c("BI/2600/75", "BI/2700/75", "EN/975/75"),
                         sera     = c("NL/172/96", "BE/172/89", "FU/172/02"))


}
\seealso{
Other {functions for working with map data}: 
\code{\link{acmap}()},
\code{\link{as.cpp}()},
\code{\link{as.json}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{removePoints}},
\code{\link{save.acmap}()},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{selectedOptimization}()}
}
\concept{{functions for working with map data}}
