% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_subset.R
\name{subsetMap}
\alias{subsetMap}
\title{Subset an antigenic map}
\usage{
subsetMap(map, antigens = TRUE, sera = TRUE)
}
\arguments{
\item{map}{The antigenic map object}

\item{antigens}{Antigens to keep, defaults to all.}

\item{sera}{Sera to keep, defaults to all.}
}
\value{
Returns a new antigenic map containing only match antigens and sera
}
\description{
Subset an antigenic map to contain only specified antigens and sera
}
\seealso{
Other functions for working with map data: 
\code{\link{acmap}()},
\code{\link{addOptimization}()},
\code{\link{agReactivityAdjustments}()},
\code{\link{as.json}()},
\code{\link{edit_agNames}()},
\code{\link{edit_srNames}()},
\code{\link{keepBestOptimization}()},
\code{\link{keepSingleOptimization}()},
\code{\link{layerNames}()},
\code{\link{orderPoints}},
\code{\link{read.acmap}()},
\code{\link{read.titerTable}()},
\code{\link{removePoints}},
\code{\link{save.acmap}()},
\code{\link{save.coords}()},
\code{\link{save.titerTable}()},
\code{\link{subsetCommonPoints}}
}
\concept{functions for working with map data}
