% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stress.R
\name{tableColbases}
\alias{tableColbases}
\title{Calculate column bases for a titer table}
\usage{
tableColbases(
  titer_table,
  minimum_column_basis = "none",
  fixed_column_bases = rep(NA, ncol(titer_table))
)
}
\arguments{
\item{titer_table}{The titer table}

\item{minimum_column_basis}{The minimum column basis to assume}

\item{fixed_column_bases}{Fixed column bases to apply}
}
\value{
Returns a numeric vector of the log-converted column bases for the
table
}
\description{
For more information on column bases, what they mean and how they are
calculated see \code{vignette("intro-to-antigenic-cartography")}
}
