% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stress.R
\name{tableDistances}
\alias{tableDistances}
\title{Return calculated table distances for an acmap}
\usage{
tableDistances(map, optimization_number = 1)
}
\arguments{
\item{map}{The acmap data object}

\item{optimization_number}{The optimization data to access (defaults to the currently selected optimization)}
}
\value{
Returns a list with a matrix of numeric table distances and whether
they are associated with more than or less than titers.
}
\description{
Takes the acmap object and, assuming the column bases associated with the
currently selected or specifed optimization, returns the table distances
calculated from the titer data.
}
\seealso{
Other {map diagnostic functions}{functions relating to map stress calculation}: 
\code{\link{logtiterTable}()},
\code{\link{mapDistances}()},
\code{\link{mapResiduals}()},
\code{\link{pointStress}},
\code{\link{recalculateStress}()},
\code{\link{stressTable}()}
}
\concept{{map diagnostic functions}{functions relating to map stress calculation}}
