% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_titers.R
\name{titerTable}
\alias{titerTable}
\alias{titerTable<-}
\title{Getting and setting map titers}
\usage{
titerTable(map, .name = TRUE)

titerTable(map, .check = TRUE) <- value
}
\description{
Functions to get and set the map titer table.
}
\seealso{
Other {map attribute functions}: 
\code{\link{name}},
\code{\link{titerTableLayers}()}
}
\concept{{map attribute functions}}
