% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_props.R
\name{titerTable}
\alias{titerTable}
\alias{titerTable<-}
\title{Getting and setting map titers}
\usage{
titerTable(map)

titerTable(map) <- value
}
\arguments{
\item{map}{The acmap object}

\item{value}{A character matrix of titers to set}
}
\value{
Returns a character matrix of titers.
}
\description{
Functions to get and set the map titer table. Note that when setting the
titer table like this any titer table layer information is lost, this is
normally not a problem unless the map is a result of merging two titer tables
together previously and you then go on the merge the titers again.
}
\seealso{
\code{\link[=adjustedTiterTable]{adjustedTiterTable()}}, \code{\link[=htmlTiterTable]{htmlTiterTable()}}

Other map attribute functions: 
\code{\link{acmapAttributes}},
\code{\link{adjustedLogTiterTable}()},
\code{\link{adjustedTiterTable}()},
\code{\link{dilutionStepsize}()},
\code{\link{logtiterTableLayers}()},
\code{\link{mapDescription}()},
\code{\link{mapName}()},
\code{\link{titerTableFlat}()},
\code{\link{titerTableLayers}()}
}
\concept{map attribute functions}
