% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_methods.R
\name{view.default}
\alias{view.default}
\title{The default viewing function}
\usage{
\method{view}{default}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Arguments to pass to \code{\link{plot}}.}
}
\description{
Visualises an object using plot.
}
