% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_methods.R
\name{view.default}
\alias{view.default}
\title{Default method for viewing objects}
\usage{
\method{view}{default}(x, ...)
}
\arguments{
\item{x}{The object to view}

\item{...}{Additional arguments, passed to print.}
}
\description{
Default method for viewing objects
}
