% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_methods.R
\name{view}
\alias{view}
\title{S3 method for viewing objects}
\usage{
view(x, ...)
}
\arguments{
\item{x}{The object to view}

\item{...}{Additional arguments, not used.}
}
\description{
S3 method for viewing objects
}
\seealso{
Other {functions to view maps}: 
\code{\link{RacViewer.options}()},
\code{\link{RacViewer}()},
\code{\link{export_viewer}()},
\code{\link{ggplot.acmap}()},
\code{\link{mapGadget}()},
\code{\link{plot.acmap}()},
\code{\link{setLegend}()},
\code{\link{snapshotMap}()},
\code{\link{view.default}()}
}
\concept{{functions to view maps}}
