% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_methods.R
\name{view}
\alias{view}
\title{View objects interactively}
\usage{
view(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Objects to pass to the method}
}
\description{
Generic function for viewing objects interactively.
}
