% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_methods.R
\name{view}
\alias{view}
\title{S3 method for viewing objects}
\usage{
view(x, ...)
}
\arguments{
\item{x}{The object to view}

\item{...}{Additional arguments, not used.}
}
\description{
S3 method for viewing objects
}
